fix mp lobby autodownloading of addons
Makes the following changes:

When we download new add-ons,
- tear down the mp ui by throwing an exception
- refresh the changed game config, and reload for this game mode
- relaunch the mp lobby, without making a new network connection
- make sure to ask the mp lobby to refresh the games list, so that
we update the joinable / needs addon status appropriately.

After this commit, I have tested that I can prompt other clients
to download add-ons, and that they will do it successsfully and
then be able to join the game.

Still do: Version checking. And, check if the host has an old
version.


Former-commit-id: 9a32314761c72ad9d7dadede64439d27b44dde7f
Former-commit-id: 9a7b62ecff96922db3dac75b326388d6c264ba68
Former-commit-id: cd0998b342951c07d69a638723d79f97c40238cd
Former-commit-id: d9f50aecc587855d91b30ceab3e5134147c1bbf6 [formerly 74f89bb3e59cd98bf23dc76b758d73e83fcff6c5]
Former-commit-id: b3d78a5251dfcce957e09144be40e5f4217c2349