wmltools: refine comma_split function
First: minor change of the argument name passed to the function
from "value" to "csstring", since the string might not be an
attribute value.

Second: more testing pointed out that an empty list would not
meet the "if list:" condition and would not be extended. So
make it explicitly "!= None".

Third: when asking on IRC about trailing whitespace, it was
noted that utils::split also removed empty entries, which
reminded me that I should do that as well.

The biggest change is that the function will now rstrip() by
default, with other options for dealing with trailing
whitespace offered.
Former-commit-id: 0947013b06ba911bae081e57e0bd0f87f4fe28bb
Former-commit-id: 46c74bfb2ed3c1b6ffcad7f744aa962f21c484ec
Former-commit-id: 70aaf79a3d7259c8d44c6307148b12c959fe3119
Former-commit-id: 23ad12234b1447ea8fcac3d149af881740533bb5 [formerly 48284018b8344d7644b6fdc4bd45f9b014331438]
Former-commit-id: 824ab67e6ece52b3e87928c0971eae93d855f533