add addon version checking and reporting to multiplayer lobby
This also includes a refactor of the "set_game_items" function,
which was quite large and difficult to understand at a glance.
The refactor splits most of the work among 5 or 6 helper functions.

We also add some reporting of add-on status to the name, just as
with the (Unknown Era) string.

We also give the mp lobby a reference, at construction time, to
the list of currently installed add-ons, which will have the same
lifespan as the game_config reference which it is passed. This
ensures that this list is not changing based on add-on directory
contents, unless the lobby gets rebuilt. This is for the sanity
of the lobby, but it also makes testing easier.


Former-commit-id: 7b32dabb24ca48d13e849ae58b872a865e9477d6
Former-commit-id: a1da3f619fbbdbaec40425558a0b702ca0cc9e3f
Former-commit-id: 01ffa8b84589642e6975c527999608ef3cbce74a
Former-commit-id: 38d06690a000dde90404e046bd893d6f7517a4b7 [formerly 1258027965b59785e2c8622abd8db7e020059605]
Former-commit-id: bd31f51aff52df3aaa859959cf5211b6264f961d