fix an assertion failure if an add-on is not available on server
If someone hosts an mp game using an add-on which is not published,
the client of another player cannot know that, so it may prompt
them to connect to the add-on server and try to download.

However the previous implementation of "ad_hoc_addon_fetch_session"
used the "addon_at" helper function, which gives an assertion
failure if the add-on does not exist, presumably because this was
used only with the add-on manager gui before.

In this commit we change the fetch session function to explicitly
search for the addon id and handle a missing add-on with an error
message rather than assertion failure.


Former-commit-id: 9119583230ca8398b64ee6427061040b73b8b846
Former-commit-id: 6090b208574fdc1f5d785e54fe431fe20ad4a74a
Former-commit-id: 3de39a31c87a598fcc45baefd9eebf88476020ae
Former-commit-id: 59180a8e971fedfec8b1337070484b88c6dee6f4 [formerly 06eb2f807a18c924c7cba23f16464c4f5564499d]
Former-commit-id: d9454c06f6b80d6d6aa31b6c957eab0c3371d1be