Add a newline between objective sections
This is set up that there will be no newlines at the end of any section, unless you have
only a summary key, which no one would do. However, there will be a newline at the top of
the dialog if you don't have a win objective, but no one would do that either.

Note that because of newlines not showing up properly on Windows, this won't be viable
on that platform.


Former-commit-id: c34e0edc5996543fe3efd1f6f4eb4d241229d796
Former-commit-id: 3aa0df3b243b1d02eb166ca7ba13c6e8d4938305
Former-commit-id: 66b9e147adff969f9391d293e2ef18f9be02411e
Former-commit-id: 9711cd0bce50d8936225b78dd1bdc09e078d42b2 [formerly 6dedf91460749f2f5b51ce011e83963b710c3d9d]
Former-commit-id: 86ab09a39ef16d5f997d9c5c2e568e659d2d303d