allow putting units on recall list for nonpersistent sides
persistent=no does not imply that the side does not have a recall list:
1) persisent=no did prevent units from this scenario to be carried over
to a next scenario but it did not prevent units from a previous scenario
to be carried over to this scenario.
2) [unit] in [side] and [unit] in action wml can give recall units to
nonpersistent sides.
3) The wiki does not mention that a side with persistent=no cannot have
recall units.

So i removed checks for [side]persistent= so that [unstore_unit] and
wesnoth.put_recall_unit can put units to the recall list and also tidied
up some other code that assumed sides with persistent=no to not have a
recall list.


Former-commit-id: 9faab217185063bf55a9da04f5bc84c334c7fa05
Former-commit-id: 0aee20b68feb68b28cbc4c0282a04898525dfe3a
Former-commit-id: cef06130b204a3cf7acd3326cd97d07d26ff860e
Former-commit-id: f1dd3bcddbd999bdf007de99ba0334db553e495a [formerly af4d81ec3504b4e6d56b507355aebc5399b2b017]
Former-commit-id: f724c103ee3928ecc881ac56e52cdb27e6be0345