Herding Micro AI: bug fix for dogs sometimes having moves left at end
The problem is due to a dog adjacent to a sheep being excluded from the
dog_move CA (so that it can act further later on in case the sheep
moves again). It is fixed by adding another CA at the end that simply
takes all moves away from all dogs once all other CAs are done.


Former-commit-id: 109d42c0f0c026fb2a00638778a7cadbd3e879c9
Former-commit-id: 56a356de1d15612420a8685c6082bf6a7add68ea
Former-commit-id: 8ee1a14682b8aa3867ed3189911933205c9ac6d1
Former-commit-id: 9bb7e3185a1e809d31fffca907b15d85805b0961 [formerly 3f5e2c05c8b748c99056e3b389bc2aa57797bee1]
Former-commit-id: 05e1fa5a51a4fe27681954bfb3720ba2bb480f16