Herding Micro AI: bug fix for dogs sometimes having moves left at end
The problem is due to a dog adjacent to a sheep being excluded from the
dog_move CA (so that it can act further later on in case the sheep
moves again). It is fixed by adding another CA at the end that simply
takes all moves away from all dogs once all other CAs are done.


Former-commit-id: abc0fb3ced20380d6ce8eece4d338fe39c68bcaf
Former-commit-id: 1bbed3b4036cebb4a2291f12f5623ecf72878c2f
Former-commit-id: c68da43c5b628daf65346402bea922563284d95c
Former-commit-id: acd8d525d1230ae3f905cdf68c82929b889322ce [formerly 86bc3bd859555dc7488073966a55a730bc226b08]
Former-commit-id: e1a991e3fbb4fc98f2532184862bd06326289921