Remove superfluous 'sane' flag from version_info
This also removes the accompanying version_info::not_sane_exception
exception type, as well as the version_info::good() method.

The flag is never naturally set to false during version_info
construction/parsing and actual invalid version numbers are always some
variation of "0.0.0foo" once reserialized. The few cases in code where
the flag is forced to be false can be trivially adapted to not require
its existence. In effect, the associated exception was never thrown and
the good() method always returned true, save for the aforementioned
exceptions.

All in all, this was a preposterous waste of time and indentation
levels.


Former-commit-id: 7029715be035500d72f707b00509b71ce19f3a6d
Former-commit-id: c0dfca56207ad453620e42c4e0a6399afbb5e078
Former-commit-id: 606970d6a02fcb9e31d21d02a9514999eb342c67
Former-commit-id: e39f9024d50a04d7ddb5f95b1a1dc9c5b0ffe7ec [formerly 9d4df4571b075889b56b244c3b7aea863500950a]
Former-commit-id: 722791a703a241c4dbe22d798113ae72fef5e0ae