fix assertion failure in tree_view gui2 widget
which could happen when folding a node with a folded subnote in a
treeview with a vertical scrollbar.

get_unfolded_size() is used calculate the size differnce when the node
is (un)folded and should so not return the size 'as if all child nodes
were unfolded' just the unfolded size of the current node.


Former-commit-id: 98f0310d0c19263a47c963081be7984d9adb3218
Former-commit-id: b7f459b05017357376c0197286f67f8ce8aec152
Former-commit-id: 263041b5dba16faf630f28dbc9c02a631c9385d2
Former-commit-id: 018394656c4eb2ac0a93eb5aa7726c54263a21a2 [formerly ca3dbe84882d0221b7ea4410f84198f4226fe00a]
Former-commit-id: 071b0cfa97983b7b9bc221ec4dc3ad2a18d38b84