suppress false positive memory leak reported by valgrind
This is basically the same as a patch I submitted to anura.

The problem is that a standard vector of pointers at static scope
is used to hold a bunch of heap allocated strings. At program
termination these are not deleted so they are reported as a leak.
We add a crude manager object to delete them and suppress the
error reports.


Former-commit-id: c42a0c7f55f8ba15f454d3ef2311fce0c5990eb7
Former-commit-id: 945eb4371333e30a87730584e9f885a7e16891e2
Former-commit-id: 85054d021d3941ceab088fac1fadb16c9da6b31d
Former-commit-id: a2f053f9d0d8385d47d3c45894cf530136541b4f [formerly d83ea4286e3f45d9f1d491bff2c3b8f633f2e506]
Former-commit-id: e6d9184b357a4ee4c3e076e47c97bae9e9acc1bb