gui2/tscreenshot_notification: Don't obtain file size from caller
The caller's provided size is notoriously unreliable right now --
especially after the addition of PNG support -- because it's obtained
from display::screenshot(), which just does some math voodoo to
calculate the *expected* size of a BMP file and sticks to that.

stat-ing a file is cheap, and simplifies our constructor's signature.
This also fixes PNG files having the wrong size reported.


Former-commit-id: 7957aec5ca67c859bf2a419e7c4f7c3f8cadb8e8
Former-commit-id: df9d89f8b6eb82c198903a10d039b4dbdfd93061
Former-commit-id: 73a02d25fc7e0edbe2e9a88fd32c3f8cc86651eb
Former-commit-id: f4373cdabe0b3df272bd3715248d28b5e0cdd59b [formerly 7c9ceb9d2af00e5728cd4f8f801a2bbce10af816]
Former-commit-id: e4d2cddf81f32c33975203bb952eab4754266fc8