ui: Check whether screenshot creation succeeded and report otherwise
This displays an actual error message in the UI if screenshot creation
fails, instead of just saying "Screenshot Done" and assuming the best
(???).

NOTE: I could make it so display::screenshot() returns an object
including both the operation's result and the SDL_GetError text if
applicable, but I don't see the point for complicating this code further
right now, since it's relatively rare for screenshot creation to fail.
Furthermore, SDL_GetError's text is not translatable, which is a major
inconvenience for us.

The current error message is horribly general in scope and may not
accurately represent the actual issue at hand, but I think it's better
than asking users to check "the log" (which basic users don't really
know about) and read text that they might not even understand if they
don't know English. Someone who really wants to know what's going on
will ask us for help and we'll tell them how to access the log with
details anyway.


Former-commit-id: 91c64b35742dc93031a856d198402c3b69d95a69
Former-commit-id: e182ef052e782af619dc5e8c39a67300eb136ee7
Former-commit-id: 094d0f6ddbc8907fa362b13bf38d903fcefa1727
Former-commit-id: 82c25eaff5455af2ee65eab7c138fea9feb630c7 [formerly 3249d6d44b471792c52d4f64fd8e020a37b0f94a]
Former-commit-id: 7553a344ce21fa0b8bbc957fbd216a1d64f673f4