change member function registration to use templates and extraspace
This is used instead of the boost::bind and boost::function method
introduced before. It's a bit simpler than the boost::function code
in lua_cpp_function, and a bit faster / less can go wrong when we
don't use boost::fucntion.

It also fixes a potential future use-case -- if a lua state is
copied, but many of the functions in the environment are made with
boost bind, they become bad pointers since they point to the old
lua state, and there is no easy way to retrieve them. Using lua
extra space instead of bind ties the lua_State more tightly to our
C++ representation of it.


Former-commit-id: 80b5398460716c7e409a1ed90c11fb0012797330
Former-commit-id: 4248e49cc76605305582a2fa22b1d368ad3b58aa
Former-commit-id: 3f1aa5c1c305ca6882bdc9c2f63233b1e8c67d9c
Former-commit-id: f560c595fb44a946913f3f1744fd743eb1992f99 [formerly c5f27b37eb2a25b2f4928eff9bf1345378f691b1]
Former-commit-id: e48ce00fff1eb48208423ff147826fa9a398cf6b