remove most OpenPandora specific items
The only thing that the ENABLE_PANDORA flag does now is to allow the
resolution of 800x480 and using a special theme automatically.
Everything else is 100% "mainline code". This means if someone wants to
package for pandora several system libraries will have to be
handcompiled and shipped in the pnd including:
* >=boost 1.48 (at least one component (locale) requires this)
* >=SDL_mixer 1.2.12
* VorbisFile


Former-commit-id: 5d8862457fb947e3f9cf7a0e453164b5b6dfb703
Former-commit-id: 47852d5d5d8f1a5f55b8905b5b14cb14262d933d
Former-commit-id: 31de6218779d479ee28368aa6035c1fe9bab9742
Former-commit-id: 92cab5275ae8599ce2e794bc41a585ea476b5213 [formerly 7fb238c1da89a3255432f1cd88095fddee08f238]
Former-commit-id: 6815f95d4f653766a40ef84866d74a9ade115ec5