Add --wconsole for allocating a Windows console on startup
This is only recognized by the game client at the moment. The
implementation is a bit hacky but it seems to work. Thanks, SDLmain.

The cwesnoth.cmd Windows batch file passes --wconsole to Wesnoth in the
command line, along with any other arguments passed to it. This is
intended to be *the* use case for --wconsole, since otherwise the fact
that Windows won't allocate a console for us on process initialization
(due to Wesnoth being compiled with the GUI application flag) causes
issues with the Command Prompt shell -- namely, the shell returns
before Wesnoth allocates its console, allowing shell input to happen
concurrently with Wesnoth's output to console.

cwesnoth.cmd should be used in shortcuts and given Wesnoth's install
path as its initial working directory.


Former-commit-id: c906dcef34855f4a6631e5fab6c3c288824e1127
Former-commit-id: 5f261836c991dfb0011b80a8e31dc47d823cebbf
Former-commit-id: 9b6ab7b276a87c5dc73cb6b4007d7c6818cabb5a
Former-commit-id: 790821375cd7aa78d7737be8a64b8b79266f0281 [formerly 2d1d792ecbab53d8c803b01fb447a69c9a6513a8]
Former-commit-id: c28c75d8aeafb1fcf30e310aba66a98657c3b964