wmliterator: fix unterminated loop in lua-strip
An old add-on triggered a wmliterator crash with this comment:

    #>>>> !!!!! REMOVE THIS AFTER TEST !!!!! <<<<#

The traceback showed that the crash came from the lua-stripping
code, which interprets "<<" as the start of a lua string. But below
it is code to remove quoted strings, and it doesn't crash, even
though there are cases where authors forgot to close a quote.

Two key differences stood out in the otherwise similar second
code: only looking for the endquote string in the text after
beginquote, and testing that endquote was less than 0, not -1.

Changing both gets the loop to terminate. Making it search the
text only after "beginquote+2" means that ">>" will no longer
be found, giving endquote a value of -1. But -1 is not less than
-1, so that must be changed to " < 0" to close the loop.
Former-commit-id: 584ee052b3975fbcdbd31f03f22fce2a989ade3e
Former-commit-id: 231f1a94987c9534adcaab3d76e08336f297399b
Former-commit-id: 8ccef37c28b5d050ff94c0a171e59de6299082ee
Former-commit-id: 4bfb7fb4f19e2518e983efe0abbb9df2008628a7 [formerly 54edbe55024d977fd6bc28464aafa9629d8454ef]
Former-commit-id: eaa04969e3901117565712b454c9e2f633f90b63