wmliterator: fix unterminated loop in lua-strip
An old add-on triggered a wmliterator crash with this comment:

    #>>>> !!!!! REMOVE THIS AFTER TEST !!!!! <<<<#

The traceback showed that the crash came from the lua-stripping
code, which interprets "<<" as the start of a lua string. But below
it is code to remove quoted strings, and it doesn't crash, even
though there are cases where authors forgot to close a quote.

Two key differences stood out in the otherwise similar second
code: only looking for the endquote string in the text after
beginquote, and testing that endquote was less than 0, not -1.

Changing both gets the loop to terminate. Making it search the
text only after "beginquote+2" means that ">>" will no longer
be found, giving endquote a value of -1. But -1 is not less than
-1, so that must be changed to " < 0" to close the loop.
Former-commit-id: 5374a5ad05f148f5b1a6e9d8fc932d4620f885a1
Former-commit-id: 3862a79055d0433bee333121d1f8333e04ce7b1a
Former-commit-id: 026465af821c52975f32871315dd976073a4b156
Former-commit-id: fbb4a68276a90d5d76327845b48cf499f4328cd2 [formerly 522ac9e261b3f6c36ea21918e66e1ec846304ff3]
Former-commit-id: 06cf785953921e52b2de711ca2ec3aae868c1319