wmliterator: fix unterminated loop in lua-strip
An old add-on triggered a wmliterator crash with this comment:

    #>>>> !!!!! REMOVE THIS AFTER TEST !!!!! <<<<#

The traceback showed that the crash came from the lua-stripping
code, which interprets "<<" as the start of a lua string. But below
it is code to remove quoted strings, and it doesn't crash, even
though there are cases where authors forgot to close a quote.

Two key differences stood out in the otherwise similar second
code: only looking for the endquote string in the text after
beginquote, and testing that endquote was less than 0, not -1.

Changing both gets the loop to terminate. Making it search the
text only after "beginquote+2" means that ">>" will no longer
be found, giving endquote a value of -1. But -1 is not less than
-1, so that must be changed to " < 0" to close the loop.
Former-commit-id: ac7cbfb738736fdc470c786ec784104e3e0358c6
Former-commit-id: 2de4dec2c3a81f73e7ac113e3712d4ff3a2017e9
Former-commit-id: 9019a2ce094b2ed29d01979c185223d067bcdea7
Former-commit-id: 922bbf7bd5e64f09dc3b7c3ba10660de00f7dbc9 [formerly fe571877f2ba919679956ab6f09b42c4cc664201]
Former-commit-id: b1a0a5faae73807a3f656823dc70e57a98421355