clear the redo stack when toggeling shroud updates
The previous behaviour causes OOS, assume this situation:
Player1s turn, DSU is activated, player1 moves  a unit and then undoes
the move. then he deactivates dsu and redoes the move. The result is
that the move did not uncover shroud (but it should since the move was
now done after DSU was deactivated).


Former-commit-id: 0bd87767a8952300d28f959126e2ed5ba2519866
Former-commit-id: 455707b964a797fb364ab6421d521bfb50c72df4
Former-commit-id: b1cac032ce251a363d30ddaed17ddcba0762b024
Former-commit-id: 411c30d7629183225884794f812978fa050242a6 [formerly 378754058488af5a54ad6c7f9ce86984d093da10]
Former-commit-id: 8e97b21f8d79d90f6611b6ca6e0e81654bf0b808