image: Use IMG_Linked_Version() to obtain runtime SDL_image version
This is a function, not a preprocessor macro, so its existence can't be
checked using #ifdef. Furthermore, I can't think of a valid reason for
the SDL devs to arbitrary drop it in a future release without warning,
so we can safely assume it's always present for now.

(Also, it's 'IMG', not 'Image'.)


Former-commit-id: 36a46c7a0761581752fd3b92ba8184f8abbcaf72
Former-commit-id: 7d66b992bdcaa98cd2a5b655cad08188a8380d3b
Former-commit-id: 34ec8afdca155a4f2c0b3d355f95096de79bff0b
Former-commit-id: 79b5bc94e9e9fe83798317a1558223f6b9e5cbd0 [formerly d7a5f34aaab2b3bdd63be5777f1cad9148b5b529]
Former-commit-id: a4e4403b8fd717723b4162322c7e5652f4127e1a