net: Use SDLNet_Linked_Version() to obtain runtime SDL_net version
This is a function, not a preprocessor macro, so its existence can't be
checked using #ifdef. Furthermore, I can't think of a valid reason for
the SDL devs to arbitrary drop it in a future release without warning,
so we can safely assume it's always present for now.

(Also, it's 'SDLNet', not 'Net'.)


Former-commit-id: 0ca19827226c7cb43221d46d3a4bb167f9c2d7d3
Former-commit-id: c1967f995e99c35fd09d588e249e16732c722d01
Former-commit-id: 2e95aa5e721bfe0bfddd07c366e2ab2032dfe0aa
Former-commit-id: f5f98fb6c553317f6dcdc5306979e3ed814c5df5 [formerly bd11918aa928bfbcfe7317b7ddf9803e677af505]
Former-commit-id: bd6cb9e24cdbaf6f456e5d3ceb32e0b89113a5dc