dont create unit ids based on fake ids.
To do that i replaced the underlying_id  integer with a struct that
holds an integer and has functions to know whether this is fake id or
not.

Fake unit ids can be created via whiteboard or during unsyced contexes
(like wesnoth.sync_choice or ai code)

This commit is intended to fix the case that soneone created a temporary
unit in wesnoth.sync_choice and then returns that units id which might
not be unique   (becasue it is based on a fake id).

This commit also removes a comment that is invalid by now because config
can hold unsigned long long.


Former-commit-id: ddca51531c09d6cc3b67ed1ba151f14860bb45b0
Former-commit-id: bd1015903cd27619576233ada65f354d84c2ef10
Former-commit-id: 2e2b53a2367b288250ccbe9f10d8c03aaf4323b5
Former-commit-id: b4eeb354d7626d47a9291fac16d2c98791d4bed7 [formerly b1aecdec84914bba83c21a5fa70c438fc58ee2aa]
Former-commit-id: 3329cf5205ad14dd87072a5b24ebfc6fb55add7d