ttext: Add support for selecting monospace font family
This selects DejaVu Sans Mono instead of whatever sans serif font is the
default for the current locale is. I'm not entirely sure what to do to
handle non-Western European languages since this is intended to be used
for displaying code or debug info anyway.

Because I can't be bothered to figure out how the codepoint fallback
works for GUI1, I'm only adding support for using this with the
Pango/Cairo ttext wrapper, which is used by GUI2. I'm also leaving it up
to fontconfig to find the correct font file for the font when selected.
Hopefully this will suffice for now.


Former-commit-id: 932332809ca47949e84718910cb29ca7df82bf57
Former-commit-id: 68d09fb00039ddd5fa47934a1c1e5eadab119d68
Former-commit-id: 4ffb5d413c4a27fe5e224cd60e55cab76bffbf18
Former-commit-id: f8bee584535f1691ffa4eb6fa08bbb4deabb8558 [formerly 4cef2c7dbb768ffb6bd0b171314ffcbdf3069dec]
Former-commit-id: 9c9b175a7acc766320b157bf818a214572b79e92