font: Provide our own fontconfig settings (bug #20337)
Due to an unfortunate combination of Cairo's use of premultiplied alpha
and our engine expecting to be able to blit pre-rasterized text surfaces
with an alpha channel into arbitrary surfaces, we can't safely use RGB
subpixel hinting without getting glyph color glitches. This is
particularly noticeable in GUI2 dialogs when the system fontconfig
settings dictate subpixel hinting to be enabled. Bug #21648 is just a
Windows-specific case of the same issue exacerbated by an as of yet
unaddressed peculiarity of Cairo's premultiplied alpha format on
Windows.

SDL_ttf avoids the issue entirely by always using an equivalent of
hintstyle=full and rgba=none with FreeType directly. There are very few
UI components in Wesnoth using SDL_ttf anymore, but they are still large
enough to make the rendering differences rather jarring (MP lobby,
preferences dialog, parts of the theme UI).

Our new custom fontconfig settings use hintstyle=full and rgba=none to
produce the same results with both SDL_ttf and Pango/Cairo render paths,
and eliminate the subpixel hinting glitches. This Works For Me™, but
fontconfig and Cairo generally seem unwilling to cooperate with me and
may produce different results on other systems.

It remains to be seen exactly how this patch impacts Apple OS X users
due to bugs #23560 and #23628. Windows users are unaffected and still
depend on the "fix" for #21648 because fontconfig is apparently disabled
for most intents and purposes on that particular platform (see also
commit 4ded0768a31b3448a2054f5d4dede1a9cc8541fb [formerly 72e2c4c6e32d122e1cdfd230cb4d47300b455c3f] [formerly 5767770025faea71aec3ee72b3b11f031ac54c27]).


Former-commit-id: e0793606dcaf7007de904d9a766b3b85e4c9a939
Former-commit-id: eca9ae0bdcfb7b95f81fbd84b2f6db3db8f72c3f
Former-commit-id: 9efbafb364df4e823be137c0096627479df20244
Former-commit-id: 890ed8e0c218f4e857bb7cfc81a9a61fc153727f [formerly 592d09110b54c02a179c56c9704d8b63035c01b8]
Former-commit-id: 0e9c7da6d400134f455f8b8c513684cecced9e66