font: Provide our own fontconfig settings (bug #20337)
Due to an unfortunate combination of Cairo's use of premultiplied alpha
and our engine expecting to be able to blit pre-rasterized text surfaces
with an alpha channel into arbitrary surfaces, we can't safely use RGB
subpixel hinting without getting glyph color glitches. This is
particularly noticeable in GUI2 dialogs when the system fontconfig
settings dictate subpixel hinting to be enabled. Bug #21648 is just a
Windows-specific case of the same issue exacerbated by an as of yet
unaddressed peculiarity of Cairo's premultiplied alpha format on
Windows.

SDL_ttf avoids the issue entirely by always using an equivalent of
hintstyle=full and rgba=none with FreeType directly. There are very few
UI components in Wesnoth using SDL_ttf anymore, but they are still large
enough to make the rendering differences rather jarring (MP lobby,
preferences dialog, parts of the theme UI).

Our new custom fontconfig settings use hintstyle=full and rgba=none to
produce the same results with both SDL_ttf and Pango/Cairo render paths,
and eliminate the subpixel hinting glitches. This Works For Me™, but
fontconfig and Cairo generally seem unwilling to cooperate with me and
may produce different results on other systems.

It remains to be seen exactly how this patch impacts Apple OS X users
due to bugs #23560 and #23628. Windows users are unaffected and still
depend on the "fix" for #21648 because fontconfig is apparently disabled
for most intents and purposes on that particular platform (see also
commit 4ded0768a31b3448a2054f5d4dede1a9cc8541fb [formerly 72e2c4c6e32d122e1cdfd230cb4d47300b455c3f] [formerly 5767770025faea71aec3ee72b3b11f031ac54c27]).


Former-commit-id: 86a378c67854119f06800d16e2e04a917acd1a80
Former-commit-id: 8f21201c442f27b2fb0cec25b26a89a3c31fd44e
Former-commit-id: 0d5a281d4577791389c08aaaa056230b2f7f0fcf
Former-commit-id: ac41edc5f8a2c85aa199550fa82ffa17d9a99748 [formerly 9cf39cae1d0a62ca170c22cbab06114454d73aeb]
Former-commit-id: 2e1fe7d6062bf1ecc5d313c23d46096a3d0a014f