add parse method and static name method to MAKE_ENUM
enum.parse("str"); tries convert "str" to an enum and sets enum to "str"
on success. Returns true on success otherwise false. This is useful
because string_to_enum always throws an exception on failure which is
slow if failue is not that unlikeley.

enumtype::name(); returns "enumname" (the name of the type). This is
intended for debugging.

Former-commit-id: 1e44fa6ec967f4aad3839854089714b77d65dc58
Former-commit-id: 6282f8976005958a76c98bc03c5768076b8359c0
Former-commit-id: 7d15f291b1521f3160f6d8abb42d780b6f64ebe2
Former-commit-id: 13aeb8490f1b0b007e478bc3e5bcd755d2a1c515 [formerly 65fefe1d723428512df8a5a2480abbeaf5a52183]
Former-commit-id: cdfbe792540e12a902a27dae6642cc813ae133d4