mp: Implement whisper_friends_only option in GUI1 lobby with a periodic warning
This option was already implemented in the GUI2 lobby (without a
warning), which resulted in a misleading advanced preferences entry.
(Forum thread: <http://r.wesnoth.org/t42423>)

The warning works as follows: you get alerted of individual clients
trying to contact you while you have this option enabled, but you'll
only get the alerts every 5 minutes. This seems like a fair compromise
between dealing with abuse and reminding people that this option is
enabled in case they accidentally did so or only intended to do so
temporarily. Also, if the sender is in the ignore list, this will not
produce a warning.


Former-commit-id: d5e4124bf184c00ae06f283b9d50c7e1665a7611
Former-commit-id: 71e0ea1c201a68ccee9b0db315aa63e546cce76a
Former-commit-id: 455672f5d44a2b1833955009ca425fa0b36a1bb4
Former-commit-id: ce82ea9c28f9ceb5e64a8232953fa5aa40b7f417 [formerly 5cdad9f4647df46e2c8d1a0549abc6718fcb554d]
Former-commit-id: 75270ca83fa276d4805289d5e61531ed9da4f07f