mp: Implement whisper_friends_only option in GUI1 lobby with a periodic warning
This option was already implemented in the GUI2 lobby (without a
warning), which resulted in a misleading advanced preferences entry.
(Forum thread: <http://r.wesnoth.org/t42423>)

The warning works as follows: you get alerted of individual clients
trying to contact you while you have this option enabled, but you'll
only get the alerts every 5 minutes. This seems like a fair compromise
between dealing with abuse and reminding people that this option is
enabled in case they accidentally did so or only intended to do so
temporarily. Also, if the sender is in the ignore list, this will not
produce a warning.


Former-commit-id: 7ec3d2d2ed07df6ead001a85334af3d42792bd9f
Former-commit-id: c7e5f352cf27530aa9476a6b0dbf7f2a53f8a598
Former-commit-id: ed3813ae1b0b65d4e82c668ec0070077f4fbfb97
Former-commit-id: ef1b7a85d45d7af3c7cefae19f4856003e6b6353 [formerly 102bc9b6ffdb1b883163bd9de0d3290880082308]
Former-commit-id: efbd8427c3d83e2461c85b7b48244777620da830