veryfy that synced user commands come from the correct client
Previously it was possible that (modified) clients send user commands
like recruit or move while its not their turn. To prevent this, we now
use the from_side check (that was previously only used for dependent
commands) also for normal synced commands.

The from_side check works like this:
The sending clients adds a from_side= attribute to its commands which
contains which side issued this command.
The server veryfies that that client actually controls that side
The recieveing clients veryfy that this is the side that can currently
issue commands (the currently plaing side).

For compability with older clients  it currently just gives a
ERR_REPLAY. Later it should be changed into a replay::process_error.


Former-commit-id: db8bb0243dd6a242342ac17fd28a7a96e151a304
Former-commit-id: b5a5122e5a4642f20d7d5d22517ec7340c8871fb
Former-commit-id: f005ccc5f378962eb544ccbb01e3826e1000f0e8
Former-commit-id: 49876807d24dfa228e1a0bc540b8a6a07b348bfb [formerly 9614f7bee0c04ebd31a81c90be4dcf62159384e5]
Former-commit-id: b1676f8603e6fda0b2cb576ebc3fb47eecd43500