don't check team names on serverside
http://gna.org/bugs/?23028

previously clients could add a "team_name" to specific packages which
should only be redirect to sides in this team. But the serversided code
which calculated in which team a side was had some bugs:
1) It did not correctly handle sides that were in multiple teams
([side] team_name="team1,team2"...)
2) The serversided data about teams was not updated when wml changed the
teams, also the calculation which side was controlled by a human used
posssibly oudated information.

This commit moves this calculation to the sending client so that the
clients now must specify a list of sides to which the package should be
sended.


Former-commit-id: f2cfbc445c9180ca62df8755b87b09fa5da8d422
Former-commit-id: 0ebdd1f3e7759676f29b73af7c6e764e81d1343f
Former-commit-id: 89941858a67982068717437fe6189856b64e5993
Former-commit-id: 6bc0406b04b36b7723f5d28e608ef2cc9043b92a [formerly 0e83d0fc054abccb0df2c9952c428ca9c25d5e97]
Former-commit-id: a00c3e66b354cbdad8f402501ee8bf889a8c4958