ANL: Work around recruit lists being overridden by factions (bug #23593)
This only seems to happen when Use Map Settings is enabled. Using
faction_lock in that case forces the "Custom" faction selection, which
offers all possible leaders from all factions and doesn't override the
recruit lists.

faction_from_recruit would limit the faction selection to Loyalists in
1.10, thus ensuring all leaders are Loyalists with Use Map Settings on,
but that doesn't seem to work right now either. It shouldn't matter for
this scenario anyway as long as the recruit lists are the ones defined
here rather than the faction's.


Former-commit-id: 60baad0043b0ae1eaf209e6157dac56eac20156f
Former-commit-id: 5c4903421c279fad63044a7acfee32034599d2d7
Former-commit-id: 3fa767b45d27370d7fe68d06afd96ca2883eea43
Former-commit-id: e16f5aa4fee24e37c039fd088ef820e732c0956e [formerly 22295cc3d0bb50901aaef62e4ffdd78a10fe0374]
Former-commit-id: 1cbccf27cf9f31009fa707d555453a85f873615d