ANL: Work around recruit lists being overridden by factions (bug #23593)
This only seems to happen when Use Map Settings is enabled. Using
faction_lock in that case forces the "Custom" faction selection, which
offers all possible leaders from all factions and doesn't override the
recruit lists.

faction_from_recruit would limit the faction selection to Loyalists in
1.10, thus ensuring all leaders are Loyalists with Use Map Settings on,
but that doesn't seem to work right now either. It shouldn't matter for
this scenario anyway as long as the recruit lists are the ones defined
here rather than the faction's.


Former-commit-id: 4c276f548ca1cdb1b405ed5dfaa7e86994be5771
Former-commit-id: 36b01a5cd354d8d17e8ef170bb1fd71fc8feaa0e
Former-commit-id: 04ce9b26c2cc181df6eb3eda7017e61ac5173a7c
Former-commit-id: 031948026a35587344a15e3a9ab904f045e229b7 [formerly f332fa2454960fcd6f922f0dccaff20a5cd802b3]
Former-commit-id: 62b740fc50a9b38fa0534bbbfeb60d1b25f7829a