Fix an OOS
Previously it could happen that a client would wait for a remote user
imput while still having a non-empty undo stack. Becasue of this the
client would not send the move that issued the user input (because the
client thinks it is undoable) to the other clients which then never gave
the user input.

This resulted in a situation where the game could not proceed and had to
be aborted.

Fixed it by calling resources::undo_stack->clear() as soon as we know
that a remote user input is needed
(synced_context::set_is_simultaneously()). Also added some assertions to
guard against this situation.


Former-commit-id: e0ef143cd406fb2a393f9f961fac736d6b5c2db7
Former-commit-id: db1a083b9fbec01ef2b0c99c35d4bbd4aaf2b132
Former-commit-id: 3f311711eeb61324163a2103c00ce431f17e55d7
Former-commit-id: 5e8cd99c3f8a30ec7d4ff027d504edc8c63ddfee [formerly aa68f7ed07da1c08c2979054118045323140996e]
Former-commit-id: 9d8efc942c4c09b05d1b61cb3f0d46a5748c4abe