Fix another OOS
Previously it could happen that moves of the previous turn were not
sended yet to other players during 'side X turn' and similar events.
This caused OOS because "change_controller_wml" could be sended before
the the moves of the previous turn.

I also removed a virtual froma function that didn't need to be virtual.


Former-commit-id: a802716921578264663be6ade3be9a4cbc560a7c
Former-commit-id: f4cad78f235f4ea5903175b0eaf22d3d9569d4f0
Former-commit-id: e8ae4d4bbdd7e81bf3aa3510e6b447040b03825d
Former-commit-id: 89c07a96aaa1bd6732675e3b59b364df5050ee61 [formerly fff723f8816ca881e7b2228af96a89505980ed47]
Former-commit-id: aaf262a88ceb1c20627b95958324950aac4e822c