dont make 4mp leader quick in campaigns by default
making 4mp leader automaticly quick is unexpected for a campaigns
designer. And the intention is make it easier to port sp campaigns to
mp.
To do this i ported the quick4 mp leader code to lua.

I also moved the lua code from eras.cfg to a new file eras.lua, this
makes is easier for an editor to detect lua syntax highlichting
automaticly.


Former-commit-id: cc6674debcbf8b189c43ed45525de29d7088ab85
Former-commit-id: 08cb6e6089c1547930913f8e2c6782128a20346a
Former-commit-id: d266f6a2db9f24f442f9f72e23c630011935a7f2
Former-commit-id: 3c56cce657436aad8477b6040646f48f76584af3 [formerly c6f9177d4ceadea8745bbd453852b2d092bb2d3d]
Former-commit-id: 529feb1d3547e506ade95696ea190512d3f33945