add previous_save_id= in [side]
A feature for mp campaigns. previous_save_id lets you specify which user
of teh previous scenario should control this side. For example if
previous_save_id="Konrad" than the user that controlled the side with
the save_id "Konrad" in the previous scenario should control this side.

This also replaces the controller=<number> feature which had similar
intentions but was less flexible.


Former-commit-id: 26971022169b90aada376d058ea30147bd877354
Former-commit-id: 1020539ca8b830b097de8f8fbc44239fa8b63c81
Former-commit-id: b4a36298bcaf87446a30977ff16fba867f165475
Former-commit-id: ac4fce5550e973974081cb66a86d151dab9837fa [formerly 54907891309ef099cbaf60e40df30dabea0eab2d]
Former-commit-id: 92c9951d8fc1330692940f00386a8c89ea6a50ae