fix [modification]type=key
Previous it couldhappens that create_engines and depcheck::manager s lis
of mods could get out of sync which caused wrong modification to be
active.

Also we add a type= key to [era]s chiech defaults to mp so that only the
default era is avaiable in sp by default (for technical reasons there
must be at least one era avaiable in sp.


Former-commit-id: 518c427b2906bfb9122a58529d75c7a7b801ffb4
Former-commit-id: 7e25a352132c42817a63a008dec18d7877b0b37d
Former-commit-id: 70f200407851caf3dd09e68bda18fe649dc9fffa
Former-commit-id: 22e0c824108b1ae1cb15fd3538a426a79ac32cbb [formerly 70005db81e7e7db0da1b7ad3e3ea1a2805a9c4c0]
Former-commit-id: 4315247627c3fe73ec848213805236922f7e3c65