fix gui2::tlabel
previous labels captured the mouse after clicking on it (I don't know
why). This means that you had to to do another click after selecting
something different. This was specially annyoing an toggle_panels where
you'd often click on a label.

Maybe it was related to link labels, in any case this fixed the much
more common non-link labels.


Former-commit-id: fed9318b48957e2d40a6ad6e21dea4489ba497a0
Former-commit-id: 21a1f8fd72d6277b75ca6a177b75dbd6ac3a9854
Former-commit-id: 1bf89482f73a20e093cf8f15177c2e328b35c796
Former-commit-id: 9edbff5d8761f62393411b8edbbc12408c782fe6 [formerly 6dfffbef769fcf01f414c58dc72696ec1ec2bf0f]
Former-commit-id: c194d053d319c9d15c95a34c1ce0ee6ae69bc825