use synced choices for controller changes by wml
Using synced_context::ask_server_choice has some advantages over the
previous implementation:

Instead of calculating which client controls a side separately on each
client, we now calculate that on the server and send the result to the
clients. This makes it more robust against calculation mismatches on the
clients (OOS).

The mechanism which notifies the server about controller changes is more
robust because previously only the currently active side sent that message
to the server, but now all clients send that information to the server (and
the server ignores non-first notifications). Specially this fixes the rare
case that none of the clients control the currently active client during
that event (which can for example happen if the currently active side
disconnected from the game after the action was invoked but before the side
was modified on that client.).

We automatically get non-undoability (just like for random seed requests)


Former-commit-id: 2ecd2745778c707828d45dcb38fec32c28894a60
Former-commit-id: 156e536a79f03d032ca7b0e8c50a2003a5000283
Former-commit-id: 21f4c7e3f306a81221332a624eedb7e629e2d6c3
Former-commit-id: 82b16001a0d1707806f2bc8792850197e23d28e6 [formerly f1f3dd5073a9a30e77f1d44a78fe0872c76beb83]
Former-commit-id: 6bfc2010e831a7ed329bc331854048e373f1d888