UtBS: gave new Teaching ability to Garak
The Teaching ability redistributes a unit's experience points to all
the adjacent units on the same side at the start of every turn.
In case that the result of the division isn't integer, the remaining
points are assigned giving priority to the units with lower levels and
less experience.
If no suitable unit is found, the teacher's experience goes back to zero.


Former-commit-id: e1215eade96ef68bcdbf62e5e8dd32155ee55081
Former-commit-id: 6fc7537c2d43200f1bfb9089af59068eb3b05a22
Former-commit-id: 3384c8365d77504a981d921afdc85f12f56fb3ae
Former-commit-id: 87bc00b0cc289c3fc802c137ff1f06f865ddee0e [formerly 091e1474becec8bfb85b3432a302d91af4296772]
Former-commit-id: ff42f695cd19c700efd7e83a2149bc3dcaab8958