New centralized library versions API
Registers library versions in a single structure that can be queried
later as many times as necessary (say, from the game UI) instead of
relying on recording this information in different object files in an
inconsistent fashion. Also, it provides a pretty formatter for console
usage (e.g., in --version).

libpng also gets to join in the fun.


Former-commit-id: e9dcfd37b9c8deff278bb62787285a5151d35204
Former-commit-id: f90b9d4c505d37f39688c0f287544369425dfd3f
Former-commit-id: b23c6bf1cf3e09c98772e139d4a9cd1963be3a55
Former-commit-id: 893b87d8729835d15a39e12a36b60de7700f1a82 [formerly a499d9682a13b25b394b715d4ca1de993bf5c705]
Former-commit-id: bcb306856fff92e6ded3276cd85ac80efc8971a1