gui2/tstacked_widget: Add single layer mode and layer switching
This effectively allows us to use stacked_widget as part of the
implementation of tabbed dialogs like Preferences.

The most important thing here is making sure individual layers are
selected so that they can receive events. This works fine in testing,
and it's much easier for me to take an existing widget and adapt it for
my purposes rather than build a new one from scratch.

Another important thing is not selecting the same layer twice. This
apparently deselects it instead.

Finally, existing applications like the titlescreen are unaffected.


Former-commit-id: 42d7ce97708c0a3b3a989cd0ebb678c7ea6ce810
Former-commit-id: 524ba9878c9b2a757b680a3ad82a1d890acc0d2f
Former-commit-id: 8879520a7e357f6c4bf985fe9f564f6776934216
Former-commit-id: f01187f9630719df9ea4c4881f08469792236d8c [formerly 8d6ec1473919ea74d3ce714c3f901646bb82ee40]
Former-commit-id: 9ccdc100f76ec4a3943ac54ac468340873016dec