gui2/tgame_version: Turned into a tabbed dialog, added plenty more info
The dialog now includes the game and OS version, and has three tabs with
extra information: the Game Paths tab with the existing contents
inherited from the tgame_paths dialog, the Libraries tab with a table
listing the build/runtime library versions, and the Features tab listing
the status of optional build-time features.

Had to move a lot of code around for this, obviously. Now I have to
decide again where to move the button from Preferences -> General.


Former-commit-id: e1c6ff1efa86219dac686d000077c0053ca18518
Former-commit-id: e9d711cad7feb891e7ba8af0334f29f973e2d1e8
Former-commit-id: cc6258f3e7d81a18727b75c343d77d3df97f7b0a
Former-commit-id: 90b91d487501e1597eaa5eec19a2674fb27598f8 [formerly 48ac6f8cc76d4728b19c3c518414a38c241211c9]
Former-commit-id: 07d0054fef7b9833c6fa102e1cc6fd147057e87b