gui2/ttitle_screen: Add button to display the About dialog
It's on the bottom left corner, just like the version number. It's not
possible to have both the version number label and the button on the
same layer because only widgets on the topmost stacked_widget layer
receive events, and the label can't be on there because it messes up the
layout of the button panels row.

This currently causes the game to abort on 800x480 because the layout
algorithm runs out of space (needs 800x487) due to to the top row on the
dialog's main grid taking up a fixed amount of vertical space. This will
be corrected in a separate commit.


Former-commit-id: 392489f812ac8eab4bd8930d6fe74d9b1694bb79
Former-commit-id: 3afed8b2415fe03aed090a2f28608e4480d887e8
Former-commit-id: ad1c4533adae2ec1130e5a8285f562f1eda59001
Former-commit-id: 25100292495a2d2b45ea07399197473db4c16ab2 [formerly d7bd50748a65ae8bbc774acb3deb36d65c0d9757]
Former-commit-id: f1c4e635d3ff6eb4a514ae33bd0bf2f18f965715