gui2/ttitle_screen: Reorganize grid slightly
This frees up some space for the central stacked widget so the game
doesn't crash to desktop due to requiring a vertical resolution of 487
or more to fit it with the version label and button.

Also, the layout of the version label and button is now simpler as they
are now on their own grid outside the central stacked widget. This also
happens to ensure that the label doesn't end up behind the tip of the
day panel. Whether this is a good thing or not depends on personal
opinion.

(And incidentally, the button no longer has a margin around it. This is
intentional.)


Former-commit-id: 0619c74f06982f22a828a63c7758813814bd9450
Former-commit-id: d568aa951354da64e401043e0eecb66d86c64469
Former-commit-id: a6425e7895aa56c51995f9766d08e021249fa0be
Former-commit-id: 93cb2663f9f54a2c2a906c534ce5fe537c58fc7d [formerly a583b214153995412f35dbb6d6f397103f0e698e]
Former-commit-id: 0e5064b5b3e3affc7123a44c79340acfeacab3a8