Add manifest declaring compatibility with Windows 8.1 and 10
This is needed so GetVersionEx doesn't lie to us on these versions.

MSVC++ people: I have no idea how this will interact with your building
procedure, if at all. In particular, I believe MSVC++ is supposed to
produce its own manifest declaring our dependencies on the MSVC++
runtime libraries, and it may or may not emit the compatibility
information as well.


Former-commit-id: 47964a60055e53aba5efbc07e838f5cce6a49324
Former-commit-id: 09e4911dc0c1fb8566b21424a025f9ed205f1562
Former-commit-id: fc545ee4b825f14f9c998764f9e5342c19035f78
Former-commit-id: b5a014fec7e530c293ad73387222a7e429fafb3b [formerly 072ffd4b3659572fe77c78a94535d862576f39f6]
Former-commit-id: ba312c555cae32bd023e2bdff5cb281caeed36a4