Add correct version numbers to the Windows resource file
For the numeric fields, this means we get to treat all X.Y.Z versions as
X.Y.Z.0, and there's no string suffix for those for obvious reasons. Not
that anyone cares, really.

Now we get a pretty name for our client process in the Task Manager too,
hooray!

(Also cleaned-up some of the placeholder code that was left here by
loonycyborg.)


Former-commit-id: 7e39030809b34ef780c730d0ae011c733f6d60d8
Former-commit-id: 0e5f250434b67ac1e5c87a9d9dc02c1e080e17b1
Former-commit-id: 4f4cac1c542a846cc00248b843b1949861200554
Former-commit-id: f5088043afb62fa44bff88f8b10408db8a8cc574 [formerly 5c0dd22c22bdad1ce4be20974d60361c8b3752f8]
Former-commit-id: b6a12b41a5e70cd2333d95842533043d7a30779f