help: automatically scale unit_type.profile in the absence of unit_type.small_profile
Prior to this, profile= was used in the absence of small_profile, but not scaled, requiring
the manual specification of a 'small_profile=profile~SCALE()' line in unit type files if no
unique small profile image was provided. Since it appears the help browser is the only place
that used small_profile, we move the scaling here.

The standard value of 205x205 is used.


Former-commit-id: 3b3b6c3a2919b291b1e307bfd4588ee75f88583f
Former-commit-id: 9eafac48502061adaa574382184b24f887d6d73c
Former-commit-id: 57c0854b2b8cfe3dfceeb8b002d0902324e6f2d1
Former-commit-id: 18bf4cdc15e728fe4b7e923f9c5902b93d5fcbb0 [formerly a93644c7672df322a6becb6f935550a1194f8362]
Former-commit-id: 72a8c45cb44db606fccfb46de9b6013831539f5f