tristate buttons for gui2
This commit adds the c++ code to add tristate buttons: toggle_buttons and
toggle_panels can now have an arbitrary number of states instead of just 2
(selected and not selected). The number of states is determined by the number
of [state] tags in the widgets definition.

This doesn't add tristate definitions yet, but the plan is to use tristate
buttons later for the headers of sortable lists (No-Sort, Sort-Up, Sort-Down
buttons)

TODO: figure out whether to remove the parameter from tcontrols constructor.

Former-commit-id: a3b1dcfa0fbd2baeca64f86ea8c6019bdedb438e
Former-commit-id: b7872244633d5c96462abe7cb292e28a48491805
Former-commit-id: 6e34775741649be232e94ae85de8c4ee3157fe15
Former-commit-id: 8d2f69f44732747ff86d8337bf8930a76af8ef39 [formerly 8a7b6c619ff587c8213cfec7d1f9dd5ee8e84618]
Former-commit-id: f4971b31127e14712c240a8a7f70db70c758010c