Use listboxs sorting feature in experinemtal gui2 addon_list
Also fixes the toggle panels that show the addons description and the filter
textbox in the non new-widgets codepath.

These changes mostly effect the non new-wigets codepath. The new-widgets
codepath (which already had a sorting feature) still works like before
except that it was adjusted to the changes to the cfg file.


Former-commit-id: 9221ab70682c61f9d9de28deb81ab2039b481a26
Former-commit-id: 7dc37f0f9cb1d522da6e3cad392ccb00a9dcab3e
Former-commit-id: 454bbf5e7be7a2e3f1404fbfd3396705b347c4ec
Former-commit-id: c1392fae7681ba7f0691a732b7522c37e4df8909 [formerly c85232b539881e1fc9696b27ba8d622acbb43e07]
Former-commit-id: cbc9c8120e8ff8cb2697276873e29fb5f0ee0e44