Windows Installer: Translate 'Battle for Wesnoth' and shortcuts
The title is now localized, as are the start menu shortcuts. As localized
titles can be longer than the English string, a third line was explicitly
added to the welcome page (else NSIS cuts the line in half).

The manual shortcut points to the translated manual version, if available.

Due to lack of Unicode support in NSIS 2, only translations matching the
codepage are displayed correctly.

The desktop.ini file isn't removed during uninstallation as it may contain
info on other programs, e.g. if the player chose to place the shortcuts in
'Games'.


Former-commit-id: c5985b201c543561d59d4704459db20a8b1339d9
Former-commit-id: 6ddbb8d9fb59ad19b9721a3f9564d42e89ef49cd
Former-commit-id: 6e1f2fc0e848ce82dd51ba42450e28d8aac31f4b
Former-commit-id: 339f7c6f017084e2a8bad9b0efc8e2245dc33efa [formerly 6415dc44b8434692e970543dad0c6eba259e9eb0]
Former-commit-id: e2dcbeaa46ea57e418dc0d306334f0d5664bfe65