Windows Installer: Translate 'Battle for Wesnoth' and shortcuts
The title is now localized, as are the start menu shortcuts. As localized
titles can be longer than the English string, a third line was explicitly
added to the welcome page (else NSIS cuts the line in half).

The manual shortcut points to the translated manual version, if available.

Due to lack of Unicode support in NSIS 2, only translations matching the
codepage are displayed correctly.

The desktop.ini file isn't removed during uninstallation as it may contain
info on other programs, e.g. if the player chose to place the shortcuts in
'Games'.


Former-commit-id: e36b885393064c10f2e999838062b64234f9bd87
Former-commit-id: a7d3aace75d1a5094e0148e24a855a1d4cc3cac7
Former-commit-id: 49d3de453b484af4184460c845e4a14f04cf8106
Former-commit-id: 8b494c39efc20bcabcd14794d437bc1231bfc750 [formerly a536b9bd82caf2411bd4ff407a59f5f11b31f6f3]
Former-commit-id: ef20945bb78f155cc590c24792c9491e11513a3f