Fix the graphics corruption when scrolling with SDL2
SDL2 has issues with overlapping surfaces when blitting, it seems to trigger an internal bug in SDL2. This causes the blitted surface to be incorrectly filled, causing a graphics corruption. This introduces a workaround that does a copy of the original surface before blitting, thus avoiding the bug. This behaviour seems to only occur while scrolling so it's a reasonably rare occurence.


Former-commit-id: 57a170aa9bb834343113f10a08c330d7323959b7
Former-commit-id: 390d519542736386666cf7b4fe238defb128b0c1
Former-commit-id: f0861bdf941ac4f6eea3ee14aa042e25cc70f886
Former-commit-id: d10a506083dd988da78a9685e20ab6e24123cad5 [formerly e2de578afa9e1554f69dfc657440692e5a8d50f5]
Former-commit-id: 8fea23a7f1282ee8fdb915e9d65833659bd91dd6