Xcode project: update pango modules
The previous version of the modules worked only with OS X 10.9. This
version works with both 10.10 and 10.9 (and potentially with 10.8; we
currently have no 10.8 system available for testing).

These modules do not work on OS X 10.6 and 10.7. Different methods of
building Wesnoth will be provided for those OS X versions (and
potentially for 10.8).

OS X 10.5 and earlier is no longer supported by Wesnoth 1.13 (that is
unchanged by this commit).


Former-commit-id: 13e7817778905f4a7a009d5b1190490b08071318
Former-commit-id: 17f82375c62da949b9b7c729296236b45c87811c
Former-commit-id: ff21a357ded6e63359e22f068fee5299c4a608a6
Former-commit-id: 0562e27cf0c29542d092546badb7d1eabd4bad5f [formerly f24b306e9f367d3256dc8580f0ced889dc2d4300]
Former-commit-id: 74384d9d773e10c9c4c3c1151155fbce93150156