Fix version_info ctor calls expecting the sane parameter
The sane parameter was removed in commit
5d1f986c7b4c72152def1b4289dfd685518fc8c3 [formerly 722791a703a241c4dbe22d798113ae72fef5e0ae] [formerly e39f9024d50a04d7ddb5f95b1a1dc9c5b0ffe7ec [formerly 9d4df4571b075889b56b244c3b7aea863500950a]] [formerly 606970d6a02fcb9e31d21d02a9514999eb342c67] [formerly c0dfca56207ad453620e42c4e0a6399afbb5e078] [formerly 7029715be035500d72f707b00509b71ce19f3a6d], but I did not realize that
`false` would still be implicitly cast to char and invoke the usual list
constructor.

Fix this by using the default constructor instead, which does what we
want sans removed parameter.

Note that this mistake and subsequent fix has no bearing on anything
other than code presentation because false and '\0' are the same value
and the three number values are 0 with the default ctor too.


Former-commit-id: df4248500ca4c1eb1e85a2f5d0fd925499551a3c
Former-commit-id: b0d0b7f98846c0f4e0957e0895f4f05e9dc5d105
Former-commit-id: 32a0e04ced0096caf1f9566fec48887e30befdcc
Former-commit-id: 47669269944d6380c1ee47925d25614c46dc049a [formerly 467d484f9aacd558371a4419d3e0597c3b3db09b]
Former-commit-id: faf3e1546d2a88429c6ae481d7f9a5010fa8a494