This reworks the hotkey system to rely on keycodes.
There is a new class introduced, hokey_base, that is the public
interface to the hotkey bindings. It is an abstract class that relies
on concrete classes hokey_keyboard and hotkey_mouse to do the device
specific hotkey work via the template method pattern.

There are two factory methods introduced, load_hotkey and
create_hotkey that should only be used to instantiate new objects.

This commit also changes the hotkey ":" and "+" to ";" respective "="
becausse ":" and "+" are not valid keycodes.


Former-commit-id: bd74f06d82574b09e33f98264299eddc63f50d04
Former-commit-id: e25c83e90078d64d8bacd058c3f0b9973766a47f
Former-commit-id: a22e9dcf8e6283129389b0d58f05538d85c62f52
Former-commit-id: e910a00ccd38a855860ddd7541664ecac81687ae [formerly 95fd4c7ddff078607e6bb05f01c741fd3a7c4f9b]
Former-commit-id: 3d99d1d529876236de331d1ec289bffced8c1444