Fix alpha blending issue when scrolling under SDL2
SDL2 has a workaround for a broken blitting operation, and on some
hardware accelerated systems this led to an alpha blending issue where
map regions that were frequently redrawn would grow progressively
darker until the alpha blended aspects are black. This disables
alpha-blending on the copy of the scrolled map area as no further
alpha composition should be done without a complete redraw of the
affected area.


Former-commit-id: a116f7518d96e42ce73bcb3b2006556d2c733a8b
Former-commit-id: 8ecc78e16c4a0f90e6d894762a9a6068e7ff18bc
Former-commit-id: fbd962d6d4b8b4192d10ccffff1ce4186b0e93c6
Former-commit-id: a74ae0670a3123a646addea7a443daabc963685c [formerly e52daf93278035170c28477283d66b4cf7ecce78]
Former-commit-id: 9f99c4628b4807e4cfce950cd83798ef7e3b2194