Show the non-rebindable shortcuts in preferences.
They can't be changed, but it will at least let people know that the key works for that command.
Currently they're just hard-coded. A more general solution would be possible, if there's
any chance that more non-rebindable hotkeys are needed.

This also removes Cmd+Shift+Q as a shortcut for Quit to Desktop on OSX.


Former-commit-id: c8a0306bc9a7f325dce4a85f2239717f7f9ae875
Former-commit-id: cc78c56b5ebeb1d0d20c88850b48ac1debb47354
Former-commit-id: d0c56140b3ab522de173425433336ebe9e7bd5f8
Former-commit-id: 4be94cc5db8d469468b905abb89f322a253259f0 [formerly 6f44cd7a23e034734c178ac1db7ac368d41b95f4]
Former-commit-id: 3e530d573475c03b0e3e54cd65a4e831c5532b76