Fast Micro AI: separate out attacks by AI leader
In the previous version, the leader would sometimes leave the keep for
truly suicidal attacks. This commit adds an extra layer of analysis to
evaluate whether an attack position is safe enough.

This evaluation is too expensive to be done for all units for the
purpose of this AI, or even to be done for the leader more than once,
so this is separated out into its own candidate action that is only
evaluated once everything else is done.


Former-commit-id: 0a3e14837fdc694390dfc2ff270982e9086ee6f2
Former-commit-id: 5b3351626c8826458937d911b98031ac6cde763f
Former-commit-id: efc07bd02be67dce1e689e2fc9d04a681efc58bf
Former-commit-id: 21ed40d69463a9ab5cd12c8815411bcd6417c231 [formerly a8bd33bec10f53d14dc99cc70149789000a05d55]
Former-commit-id: b313ae1dadb94edf1f49cdf5e69e630505698431