refactor playsingle_controller::before_human_turn
Also removed a raise_turn_started event at the beginning of human turns
which didn't fire at the beginning of networked turns.

I don't know what the raise_turn_started did if this actually breaks
anything feel free to read it but please add a comment on why it fires
on human turns but not on networked turns.


Former-commit-id: f6560c43f638dc5cdef3fbec42634059367b4c23
Former-commit-id: 834c3dc04ba971c2ce327f867a30a14b4f61e8e7
Former-commit-id: f9722dccff6990202264312154f669d314349204
Former-commit-id: c10cedaf4b0c696391bb6303112f40c3cbf1c6f5 [formerly 62c7d87534ed90d4344f726c261bbb7063763369]
Former-commit-id: 45684b5eb2246a4a92a8c03bdfdebd3172d1dc6a