Correct Surface Alpha Blending in SDL2
alpha.cpp was apparently written as a compatibility wrapper for SDL2, as SDL_SetAlpha() has been replaced with other functions in SDL2. As it currently stands, it neglects to set the alpha blend mode which is now required when in SDL2.

A quick check of the code indicates that in most places where SDL_SetAlpha() is called, the alpha parameter is only set as SDL_ALPHA_OPAQUE, or the flag input parameter does not include SDL_SRCALPHA. So in either case there is no alpha blending anyway and so this omission didn't make a difference anyway.

However, alpha blending *is* used in fill_rect_alpha of rect.cpp. This is most noticeable in dialogues with sorted lists, such as the hot-key bindings in the preferences, where the background colour of the headings turns opaque white instead of the alpha-blended grey that it is in an SDL1.2 build. There is at least one graphical blending issue resolved with this change as well - it's possible others may be included with this.

Former-commit-id: 85fa57bd4c13a680060e5b70428709772d2caa8b
Former-commit-id: 5ae11749e577bf999177cb2f21c4ccbdfabb299e
Former-commit-id: 0bb1ccfb86229469d64a6e133547bc79da9e6cce
Former-commit-id: 6950c83a7ae94125ae10977926991e2c51fc2763 [formerly 8a91d72c098ec90904cf1fa92942325b4f0dd9ed]
Former-commit-id: 1a11bf4352e06c9104bc22e730eff3c7b287384c