Correct Surface Alpha Blending in SDL2
alpha.cpp was apparently written as a compatibility wrapper for SDL2, as SDL_SetAlpha() has been replaced with other functions in SDL2. As it currently stands, it neglects to set the alpha blend mode which is now required when in SDL2.

A quick check of the code indicates that in most places where SDL_SetAlpha() is called, the alpha parameter is only set as SDL_ALPHA_OPAQUE, or the flag input parameter does not include SDL_SRCALPHA. So in either case there is no alpha blending anyway and so this omission didn't make a difference anyway.

However, alpha blending *is* used in fill_rect_alpha of rect.cpp. This is most noticeable in dialogues with sorted lists, such as the hot-key bindings in the preferences, where the background colour of the headings turns opaque white instead of the alpha-blended grey that it is in an SDL1.2 build. There is at least one graphical blending issue resolved with this change as well - it's possible others may be included with this.

Former-commit-id: 2b7febd48973c2b283fe57a8b731662a25131fdb
Former-commit-id: 2e754861c3dc98f41112270367be12ef05574379
Former-commit-id: 36f8361cb348ecc162f1afff22d57b5f1bce2985
Former-commit-id: d699122daa2942f2b5eeb514ec300caba9d0368e [formerly 1b56b8bf78dbf521f26d2b0048d4e544576054d5]
Former-commit-id: f3beabbb3a691a2bfdf63e27fc87b25bb0d2ff69