Resolve Bug #20036
When a map is generated, the labels for the player starting locations are correctly set. However, generating another map (or even loading a map and then generating a new one) results in the old labels remaining even after the new map has been generated and displayed.

Issue appears to be that the reset labels flag is not being set after a map generation. This change sets the reset labels flag and the subsequent perform_refresh() function then includes a reset of the map labels.

Former-commit-id: 21d2694bbe7614cd33a3aba42896040bfbf9bede
Former-commit-id: ccd4aa6918e6077c85125e99140169ed52ae46be
Former-commit-id: 96541ed3dfce43f332839e5fce5b1ed4c7fe617f
Former-commit-id: d89b338db563265004c53b0fd41f2e455a607606 [formerly c88afc3698f45e087827e3876d3e272e1c9c4612]
Former-commit-id: 7177edc28f682d51d2defb1fa1c1982e1702b519