Remove unnecessary y scrolling offset from [message] implementation
The problem this tried to solve was solved long ago, when the new GUI2
[message] dialog was introduced in 1.5.x. Formerly, [message] was
displayed at the top of the screen, so with sufficiently long messages
and a small screen resolution, the viewport could be scrolled in such a
way that the target unit was covered by the dialog.

Nowadays [message] is displayed at the bottom and the scrolling
algorithm is more or less sane, so this hack is completely unneeded.

(For reference, this was originally introduced in commit
0a097dd8c07fec3c86787f5136de0c580b39a4cd. For a reference screenshot of
what it would do back in the day, see
<http://www.wesnoth.org/images/sshots/wesnoth-1.1.9-shot3.jpg>.)

Thanks to celticminstrel and vultraz for bringing this up on IRC.


Former-commit-id: 81aa1026bc0d787ac8b41f6b3addeb0a2f269fa3
Former-commit-id: 0c5735626d134e44b50c549df01b75d1a6d3b9c2
Former-commit-id: e75e0d1dbf772546df78f0f908016be88cdf35da
Former-commit-id: d67cd667c2b33218f932a14054add78a4b7e9093 [formerly 7856f3df717a596b8ba1ba92f14be46d3ccef1ac]
Former-commit-id: cec7eec801827aaaa6dab0a9c9e4d5a7e10527d6