Remove unnecessary y scrolling offset from [message] implementation
The problem this tried to solve was solved long ago, when the new GUI2
[message] dialog was introduced in 1.5.x. Formerly, [message] was
displayed at the top of the screen, so with sufficiently long messages
and a small screen resolution, the viewport could be scrolled in such a
way that the target unit was covered by the dialog.

Nowadays [message] is displayed at the bottom and the scrolling
algorithm is more or less sane, so this hack is completely unneeded.

(For reference, this was originally introduced in commit
0a097dd8c07fec3c86787f5136de0c580b39a4cd. For a reference screenshot of
what it would do back in the day, see
<http://www.wesnoth.org/images/sshots/wesnoth-1.1.9-shot3.jpg>.)

Thanks to celticminstrel and vultraz for bringing this up on IRC.


Former-commit-id: 81312aa8b650424813514a3cb93977dd933fd9a6
Former-commit-id: 0afe8dca9469ddebdd0a878dcf0529fa63dc0945
Former-commit-id: b7db3209c25b04e2bc80650434d6d7667c9fd6ce
Former-commit-id: 4a482ea35d4f761a213b7f37c56c9c86b6fe5496 [formerly f950fc8362eded5def47150e5d9d3783a37bf5ea]
Former-commit-id: 2a4aef2b30ce446e4c247c6cedb504efbe1ff5fb