Resolve crash when unit is killed before its planned actions are completed (Bug #20071)
Attempt to resolve bug #20071 where a unit with planned actions is killed before those actions can be completed. This would otherwise result in an assertion failure or crash to desktop because sa.find_last_action_of() would be making a null pointer reference (get_unit() returns null when the unit is killed).

I have tested this with one unit killed before planned actions are completed, two units killed before planned actions are completed, and a mixture of live units with planned actions completed, live units with planned actions incomplete, and units killed before planned actions are completed. The catch is that planned actions painted on the screen get left behind after their respective units are killed off, until the mouse cursor hovers over the actions and triggers a redraw. But it is better than nothing and at the very least the game should not crash any more.

I'm not familiar with all the code, but at a high level I suppose a solution to the above is to do a check on side_actions whenever a unit is killed. (I think last_action_redraw() is only called when the mouse cursor hovers over a planned action or its unit.) I think this would be more work, in places outside of the whiteboard area of code used for action planning.


Former-commit-id: 004cf1d826f6a5764826048d9e3f85c60a28e751
Former-commit-id: 6e74c194972e8eb1eb8df876a70f8a8459ecd521
Former-commit-id: 03ec311a35535be69d908b7cd32e40e2d218a684
Former-commit-id: 06be031cc39a26ac8091f3e0e7a64d9c6dd84ef4 [formerly a371937ef048063af62fb1bc15a1cbb535497771]
Former-commit-id: a86185fe117442260d8425f35ae4193738e91c22