Avoid crash when planning moves on planned recruits (Bug #18637)
Attempt to work-around bug #18637, where planning moves on planned recruits results in a crash (because get_unit() is null when move::init() is called).

This isn't a comprehensive fix because clicking on a planned recruit will still result in the usual display feedback for planning a move. The difference here is that attempting to place a planned move will result in nothing happening when the mouse button is clicked (as opposed to the assertion failure or crash which happens currently). It is better than nothing, I believe, unless there are negative side-effects of which I am unaware of.

As per the bug report, the ideal solution is to either get planned moves on planned recruits working properly, or disallow this movement altogether.


Former-commit-id: 9c19dc64124b9eed92a353ab08f63bcad1492246
Former-commit-id: bb62b242302b64fc31d86b35deb864e4e598fd52
Former-commit-id: 3f967595af562e63023db08f29b1375b19497591
Former-commit-id: 24856bcc5218629437aa6cc8172ba72f557b229c [formerly a5752c4d85fa7ffa9cf4f7c2471df944d1db30b2]
Former-commit-id: 91fe6505613301494367abef068ebef3b67a0601