Avoid crash when planning moves on planned recruits (Bug #18637)
Attempt to work-around bug #18637, where planning moves on planned recruits results in a crash (because get_unit() is null when move::init() is called).

This isn't a comprehensive fix because clicking on a planned recruit will still result in the usual display feedback for planning a move. The difference here is that attempting to place a planned move will result in nothing happening when the mouse button is clicked (as opposed to the assertion failure or crash which happens currently). It is better than nothing, I believe, unless there are negative side-effects of which I am unaware of.

As per the bug report, the ideal solution is to either get planned moves on planned recruits working properly, or disallow this movement altogether.


Former-commit-id: 936db8e9873023817e17b00bd9b8b0a1ee075f0f
Former-commit-id: ee1ea274c58253f548b3489555217f36024fbe1d
Former-commit-id: f368cab1f7ab8b78d24c5a6fa3139b67278daf6f
Former-commit-id: 7ac527b023291ba66ac6a42b1bda3fb4641337d0 [formerly 717348b204a1456e0ca2d22ee965f505533d46a1]
Former-commit-id: d389512f0848fa2fe53981e40ceb18237e8b8853