Replace 255 with SDL_ALPHA_OPAQUE where appropriate.
This is an extension of the previous correction to alpha values in attack_prediction_display.cpp.
For clarity and compatibility, use the SDL_ALPHA_OPAQUE definition instead of assuming hard-coded 255. This is defined in both SDL 1.2 and 2.0.
Where there are mathematical operations on alpha channel which assumes a value of 255, leave alpha as hard-coded 255.
Old versions of SDL treated 255 as transparent and 0 as opaque - there are still instances where alpha is 0 in a non-transparency situation, but these seem okay. Possibly because there is no alpha blending involved.


Former-commit-id: c9fb58c96101edc8a0f2a0179e334608031dc657
Former-commit-id: 370ef98d02ae7ec17e9667fc2ed2903b119654f9
Former-commit-id: 5d81fe1b3b26a7a72e16378d3994b975a43a42b9
Former-commit-id: a9102fb72fc35de3a774e867b9b33f38e9193480 [formerly d30c3a225b646fa75b4376ec1fccd5847926ecdf]
Former-commit-id: 61a55a366dde7d3059c9d782d174d94b77cb13c3