add enable_if= to mod and era events
http://gna.org/bugs/?23721
enable_if = in [event] contains a condition that is evalulated before
the scenario is startd (specially before the [replay_start] snapshot is
taken) to check whether the event should be added to the scenario.
[event]
enable_if="$($scenario_num -1 > -1)"
[/event]
will for example add the event only if it's not the first scenario (assuming
wml sets the variable) 'scenario_num' to the current scenario number.

Former-commit-id: 565e2d4f092dab6ebf524f353c483747ed9b4956
Former-commit-id: 377d898da211292e45933eb97bb175403dd8f76f
Former-commit-id: b058968999d46586f0f4c09b68be111bf04324c9
Former-commit-id: 1b1b8ca1d36c95bdaa4c812738de0fafd333c1d0 [formerly 4a9c1b39a529c192fcf4cb4307a987dbcbd9b241]
Former-commit-id: b583e7fbd7b66f420dde433fc6c1a65dbe7c9fd4