Add an [explain] WML tag
This explains (in the log) why a WML conditional has failed (or succeeded). Currently it is fairly basic, looping through and/or/not and printing out the specific conditional tag that failed as well as, for [variable], the current content of the variable.

It's used in the WML unit tests system, but could also be useful for debugging.

This commit also fixes an issue with passing multiple -a arguments to run_wml_tests.


Former-commit-id: e699b7dd5e96bb91615433e0fcedc883a978475e
Former-commit-id: cde039011196c53fc07c7538f2e464a6cb8b9ae6
Former-commit-id: 19901ac0c7bdf4ddda093c2a49f713c725f8dfd6
Former-commit-id: 62be11225383cb61a908e0d55e525e133c76fb8d [formerly e26897abd2b57a91149091053c740895a6485109]
Former-commit-id: 8f0a8afce410267c8ddf24dfcacaa09fa5e2d352