editor: Add Recent Files menu with recently loaded/saved files
File paths are recorded into the game preferences up to a (currently
hardcoded) limit of 6 and opening or saving maps adds or bumps existing
entries to the top. We may allow users to change the MRU limit in
Advanced Preferences in a later commit.

Adds two translatable strings.

In order to fit file paths in the menu without filling up the whole
screen sideways, we use only the file names for now. Because identical
file names could prove to be an actual issue later, I intend to look
into ellipsizing paths correctly in a later step.

(Note that I'm piggybacking on the crummy submenu support we already had
in the themable UI so that this can be safely backported to 1.12. It's
decidedly not optimal usability-wise, but it'll have to do for now.)


Former-commit-id: ae8576f3c748555493a3986a96ac2f131df20364
Former-commit-id: 7f7d46d88e345991cbe0d64dc627420c385a8878
Former-commit-id: b91eb5cad2640d2cde00e70cc50cff1268c44ffc
Former-commit-id: 352064fec07ea29b4f7f5e2cad9079d93eac5dee [formerly cf2db7bf55292d316bb2535f4c1d58c4ce8fc798]
Former-commit-id: 51475162d1c071b0b5653f160eec45ec241df489