editor: Add Recent Files menu with recently loaded/saved files
File paths are recorded into the game preferences up to a (currently
hardcoded) limit of 6 and opening or saving maps adds or bumps existing
entries to the top. We may allow users to change the MRU limit in
Advanced Preferences in a later commit.

Adds two translatable strings.

In order to fit file paths in the menu without filling up the whole
screen sideways, we use only the file names for now. Because identical
file names could prove to be an actual issue later, I intend to look
into ellipsizing paths correctly in a later step.

(Note that I'm piggybacking on the crummy submenu support we already had
in the themable UI so that this can be safely backported to 1.12. It's
decidedly not optimal usability-wise, but it'll have to do for now.)


Former-commit-id: 40e69c486fa43c5ce199b190e716797d6b3eb5c9
Former-commit-id: 20ae2266dc21003561dc6f9877d6625b7398380f
Former-commit-id: 59d458dcda297f55be0c072cd47aab4540d4c581
Former-commit-id: d8ec2a5e520f6e803c29cb49f74cf45526fd01a3 [formerly e0e9c37197c85754113a45c7e4a8cbf946950ccf]
Former-commit-id: 76cfa3847f02d7ac3c8b247cd5157d3d6c3b41b2