Fix Bug #23908 - Crash on resize with SDL2.
SDL2 uses shared memory to communicate with the graphics system when
using a software renderer. A resize event will cause SDL2 to
invalidate the SDL_surface that relies on shared memory and any
subsequent calls to will be SDL_GetWindowSurface will cause the shared
memory do be unmapped as the old surface is released.  This is
problematic when it occurs during a rendering cycle, because the
software rendered may also invoke SDL_GetWindowSurface, making any
references to that surface that the game contains become stale. This
will in turn cause a segmentation fault during blitting.


Former-commit-id: 45320877f1b391e899a8fb2df012bc423b9b61fe
Former-commit-id: 15d3325f378033a0a7684461b9aa42ae137fac76
Former-commit-id: feb2ffb6972adce1d3773c18ba28e33e2b220ce2
Former-commit-id: 9acc5bcb30268c5a4b444036a2a55bf3baa2ddc7 [formerly 37756b4838dcf9860da8151479272de046b3280b]
Former-commit-id: 59a4a0501f512840a8fce34fa06aceb35bd46b43