Tab-completion added for in-game whispers.
In the game lobby you can tab-complete ANY nick, both from the lobby or from a match running. While playing you can only tab-complete ingame players and observers.
This patch works in the match, it helps answering whispers from people in lobby or in other games.
The nick will be added to a string (in the same way as observers already work) and get recalled with TAB key.


Former-commit-id: f8bfd0e92e11fc6c62f69cb6d572c599af7b2bb5
Former-commit-id: 2691d09144368a1ff55b4029508bfbde396c98b5
Former-commit-id: 89ab23e18008434d73ed61519c94e87349c53913
Former-commit-id: 95e53f9f57f88477b90ba3ae7fb2fe3f1fa436d5 [formerly bceec23796061faf364b986a6bceb4eeae735923]
Former-commit-id: f9fd1880f32df23befb2c98bd3519d148cc0fae2