Avoid polluting preferences file with null commands (Bug #21969)
Only save non-default hot-keys to the preferences file if the hot-key is active (ie command != 'null').

Clearing hot-keys in memory (via the hot-key functions in the game) only sets the command to 'null', which currently gets written to the preferences file if it is a custom, ie non-default, hot-key. Over time, creating and clearing hot-keys results in a lot of redundant hot-key entries. The nice thing about this update is that it will remove redundant command='null' entries for existing preferences files as well as prevent it from being included in future ones..

I am making this change against the sdl2 branch instead of master because the sdl2 branch already contains a lot of rewriting of the hot-key functionality. Testing against master was also hopeless because hot-key handling is still rather broken building against SDL 1.2.


Former-commit-id: b668b603836c207d81e1a0bed50662de2d12d841
Former-commit-id: 1515b2b428173e562a6c7c15bc0b55f70d4236fc
Former-commit-id: d92d7af3fa605bd7f663205db42996b1cc1b7de1
Former-commit-id: bfb149436397c469c2e84c1e2b488535e39b4ba7 [formerly c1ed44bd7d77447774420b3cf4a3bc6439af77e2]
Former-commit-id: a14510ecf13d99e3a15e1d4dc227c120d7f13060