fix preferences dialog for loadgame hotkey
Some hotkeys (like the loadgame hotkey) are present in some, but not all hotkey scopes and teh hotkey preferences dialog was unable to handle that. One of the results was that it was not possible to assign the same hotkey to the loadgame hotkey and the loadmap hotkey, this commit attempts to fix that.

Former-commit-id: 43058d8576e68e2dea0fad976dbf69a870fba184
Former-commit-id: 691bfb799a76c118afdf49d10f1b3b5efe5b0da1
Former-commit-id: 927f28b7fcb1336ec2120943d16144de213ecec1
Former-commit-id: 32dba44eebf2234b724cf624d9e54b95d2ffcaea [formerly 99b91452acc64589d43b88c49abb934b909dc6f3]
Former-commit-id: 56d7da4b6809f52a8b71e185ac808c18828a2ca6