Fix bug #23908 - Crash in SDL_BlitSurface caused by integer underflow
The font rendering code under SDL2 seems to deal poorly with silly
values being passed to it as a cause of integer over/underflows. This
commit changes the types used in gui1 from unsigned to signed for
width and height. This causes a number of statements that previously
inadvertedly did unsigned arithmetics without over/underflow checking
to be doing signed arithmetics instead, which is probably what the
author intended.


Former-commit-id: eb9e3a22a42b9e4a8ec939b8aee427f064c93166
Former-commit-id: 237e2b8c15b8a0659d8cd674859ca28a2fba5337
Former-commit-id: a148d847393800a8219b3a79604152d5f772d847
Former-commit-id: a7022c238a4c4947ef035dd1240dc1bf957fbd3e [formerly b49a874d50dfdd3e3a8300930e23f52d5f2982be]
Former-commit-id: 080d74d5ae0c8c5443ba8abda1347c3e2c6ac005