log/windows: Integrate implementation of --wconsole
This makes it so we no longer try to steal the console back to a log
file when using --wconsole with the new redirection code. Now the
--wconsole switch triggers a special mode of the log file manager that
uses a native console instead of log files.

As a necessary bonus to appease compilers, the GUI2 version info dialog
now uses the correct log file path when not started with the --wconsole
switch. Yay!


Former-commit-id: af667cb810845eb04a5c72fa7d2bff25c6602dea
Former-commit-id: 73d1ee1d6ffb271b892cddb8c55a5ecbd6a415b4
Former-commit-id: abefcae80e125899f97577421285bc0136ad35fc
Former-commit-id: f580a8f1530e089adbeb37839018947d036df15d [formerly 4b519fa47ea40e3b27fcafc524540ab428b1e6fb]
Former-commit-id: c434806a40f53530fd00e65bc8d4a13cc5ef18b5