Make [*time_area] Lua API functions take a single time area id
This means the Lua code implementing the WML tags are now solely
responsible for parsing the comma-separated id list. It just so happens
that I added code for that without realizing that I was duplicating
existing functionality on the C++ side.

As a result:
 * wesnoth.add_time_area and wesnoth.remove_time_area can operate on
   time area ids that contain commas (if someone feels this is needed
   for some weird reason)
 * [time_area] no longer warns about adding time areas with commas in
   their ids (actually adding the first id of the list instead)
 * [time_area] remove=yes and [remove_time_area] still handle
   comma-separated lists of ids, in the Lua side instead of C++.

This makes the Lua API for this consistent with [event]/[remove_event].

Since the Lua API functions in question are undocumented at this moment,
it's unlikely this impacts anyone, really.


Former-commit-id: be04a46cde0e4773c53f0ca5722e75f33a89e6aa
Former-commit-id: 9e11fb4bd1c43b093a0e51cca40e09d9bd9bb7cb
Former-commit-id: 0535bddbdb31ceecbff6c786bada59a0ae018270
Former-commit-id: b4d44a625e73d25b79b0be14af1a1e8592fde51c [formerly b3c74278163f132b822bcc0da54b77b113745181]
Former-commit-id: 83d1631571f25453fb2ba40543f1f62d6e5db38b