Peek for redraw events after each call to SDL_PollEvent.
Since SDL_PollEvent internally pumps events there is is a chance that there will be a redraw event added to the SDL event queue, which could cause a crash if there is a game DRAW event before it in the event queue. This should fix it, but it's not the prettiest of fix and should be rewritten if it works to instead put draw events to the start of the internal event queue after it's been drained.


Former-commit-id: b32cc90b8a8f9988c1c6aba965f0c94db38b91ea
Former-commit-id: 36d040ae419180d1b3b59ab87d08518504199d21
Former-commit-id: c5dc9946d063143aef9d1bfa25bc68557489da9c
Former-commit-id: f81ca9964f5c1910c3955e9e49c0423fa3a73bf2 [formerly d74302918dd2e3d05fcfdea7b018e429cd57a315]
Former-commit-id: 0b60e1b5f9e01a51c90fc697b05cf6847eafbdf8