Fixes bug #24138 and disables SDL_PumpEvent() in grid.cpp
This removes the usage of the display::get_singleton() in the event
handling code since it may cause incorrect behaviour (crash) when the
singleton has been incorrectly set to null.

This also removes SDL_PumpEvents() from grid.cpp the rendering code
should not be directly interacting with the event handling and it's
causing instabilities.


Former-commit-id: 054063f0e19dbc128b8cea5215b8909864ca4d62
Former-commit-id: 5ec3d24e58b43af19ac7b669d3dfc361c8523c3d
Former-commit-id: 5dc4c47d6679212cd8ac613070e74208a6f8cc21
Former-commit-id: 7b2c158798521ef9900fc22092188c69831a5b80 [formerly df44e34211f447f4c50e0315bfd5d4c5d76b020e]
Former-commit-id: 12b79ec2b64ed721bf1e6d26e060682fcb09d6bc