Bump version
Yes, RC_VERSION_REVISION is 3 rather than 2. Because the Win32 version
info resource doesn't take text suffixes in the field where this value
is used, and a fourth number doesn't make sense for our usage, from now
on RC_VERSION_REVISION is one release ahead in order to clearly
distinguish previous release builds from in-dev builds, which are for
this purpose equivalent to the next/future release builds.


Former-commit-id: f5dd5036b8638cae4d3daa20dc4720e0dbaa687d
Former-commit-id: 6b2112a33c4dc906843342df97e1d27c2cbda3b3
Former-commit-id: dcd0269b339a321b8b351d7a029027395de81ce8
Former-commit-id: 4e61e2ead403158bfe37877e397de962e996273e [formerly 138df7e5d6761149216da20e1b7b2e2b55707ba1]
Former-commit-id: aefe30dbb20645bccd0083aafc8286e79d3162d8