Work around crash/assertion caused by shortening filenames in gui1
Theis introduces the get_max_height and get_max_width methods in gui1
menu elements. The get_max_width method is then used to check against
the magic number -1 which is used to specify that there is no max
width for the menu.
If there is a max width then the old way of doing
things will be invoked.


Former-commit-id: 4d83d7e6ca437d04d44adf98311f1a8968059e31
Former-commit-id: 75f15aca0ee82bd2604ba728aee3734a6bc5e7ac
Former-commit-id: 2ef547ed740f9459caf015507691272eaaf587af
Former-commit-id: 6c1d2d9b82281b28a26cf39065f1cefd16aa8626 [formerly 62dc9017d546ab000400389306000d02d1a59a36]
Former-commit-id: 3d7412fc4586d71e06d1361d1f3e16c615d8a13e