improve formatting in lua code.
This also fixes a bug in [harm_unit] where a second arument was passed
to wml_actions.animate_unit which was obviously meant to be part of the
first argument.

This also changes harm_unit to use wml_actions.animate_unit instead of
wesnoth.animate_unit becasue the former is better known an thus easier
to understand.


Former-commit-id: 29563888680fe2ae5c86da2b25b1b8e83f153052
Former-commit-id: 25644b1fe56c64dfcdca83a28f73763929b8be67
Former-commit-id: 3966904eeb09c32548360fefdd344f63bb0e6fd3
Former-commit-id: 0705f8811b82a974a321e44707da6e68c78d3fc2 [formerly aab05b098260ee654c946688b920e4baf7a00989]
Former-commit-id: cc991143e56705789d8987c3095d5edc8c1b509c