use enums instead of strings for controllers in wesnothd::game
This is not only faster, it is also easier to understand, becasue it is
now obvious that every enement of side_controllers_ can only be 'human',
'ai' or 'null'.

This also makes enum::parse() in MAKE_ENUM a template. This way it also works
with other string classes, specially string_span from simple_wml.


Former-commit-id: 49bc5b57f40fb17088d2646f235577bda52c24ce
Former-commit-id: db991669ad7334ce3d8cfc638fe311b8a2169fd6
Former-commit-id: dc2118774d5387e6a78ad1712a4ff63a4b8c9da6
Former-commit-id: 75f84c03c9c2672cfea2d9983bd99f9f6acdcf89 [formerly d94533c01cccfd9ddce4895b48c8394a67646cad]
Former-commit-id: 6ab7f78fd3e7ea27f0c78818aca753fff0b015f5