SDL2: Set a minimum window size
This sets the minimum window size properly, allowing SDL to hint to
the window manager that the window should not be allowed to be changed
into a size smaller than the set size. This removes the need to
programatically try set the window size on a resize event if the new
size is smaller than the minimum size.


Former-commit-id: 372acff3eb2a3b4f182c651c73cf5e165ba6af27
Former-commit-id: e82b5c210e66ee5f4a0af5a2d50655addce4f419
Former-commit-id: f567b1118ae8c2e96e53b46c17496e1d114982c4
Former-commit-id: f4e7d464a39e3439504ff3adaf1636561a97a5d5 [formerly 253a3807dd66cd0286dab4b108ecb35420aea56e]
Former-commit-id: 059ab8a519a9b290ac36a7d88a698d011f3ab86f