Fix bug #23934: Add event filtering on resize events
This adds event filtering and a timeout on actions after a resize
event. All DRAW, TOOLTIP and window events are filtered for 100ms
after a resize event has been received. The last received resize event
is injected into the event handling code when the next event is
received after the timeout has expired. This ensures that no more
resize events are dropped and that it's a better user experience.


Former-commit-id: 4bd9b3685129c98fad53ba116e618f84df0c7b5a
Former-commit-id: f320b9fe26392797688e4650d82c868bf598c5e2
Former-commit-id: bf41fec25f61cd3a905b2b90925c1648794fb97c
Former-commit-id: b65f6e9352d9fe2956aa5e40bea0c67fc5d85698 [formerly 36c26cb05c0fb2d12ac2bb61e5e75ff8e7aa1057]
Former-commit-id: f9e1154f459ae1aa5fac7e179d9c1030da8b379b