Fix bug #24209 and add a global event scope
This adds a global event scope that an sdl_handler can register to in
order to receive all current events, regardless of the current
scope. Care should be taken to not be registerd in a local event
context and the global context at the same time as it will lead to the
event handler potentially being invoked twice.

This also adds the workaround needed for #24209 in the new event
handler in the video class.


Former-commit-id: 762057039caf4ef352ee921d0a04f0d3ff26558b
Former-commit-id: 496c2dd54d0e83c6088f5ef6140d9806d5e62ab7
Former-commit-id: 7655471bf0c5c3483f8a84a82a0ef769854e91f7
Former-commit-id: 3d733999ca4ec0360af900b03a8010a50b756510 [formerly c17abd37c6200c982481fb18bde29791e034bfad]
Former-commit-id: 37a65050c70e657538da1be7fb14f289f6c43360