Fixed crash on startup on Windows in paths with spaces
If OMP_WAIT_POLICY is not set, the game relaunches itself using
_execv() after setting the environment variable. It passes the entire
argv array as argv for the new launch, including argv[0] (program
path). If the program path contains spaces and is not surrounded in
quotes, Windows splits the path on quotes and creates a separate
parameter for each part (e.g.
"I:\Battle for Wesnoth\wesnoth\wesnoth.exe" becomes {"I:\Battle",
"for", "Wesnoth\wesnoth\wesnoth.exe"}). Boost::program_options
interprets the last two as positional parameters (the first one is
skipped, as the game expects the first parameter to store the program
path) and crashes the game because only one positional parameter is
allowed.

Fixed by quoting the path in the argv array.


Former-commit-id: b8eab8b68a0e4deeaee24b7c0f1484dabcf7718f
Former-commit-id: 2f09d48de74cd26a5e3fa9424b615d83c10f2886
Former-commit-id: 52cb09836f390449f9eb5bf161da063480cf9fa3
Former-commit-id: 8e413f0910f717b097e712271b596bf7ce2d40e2 [formerly 146e1241be16b204dadb95cbfcf9f24c596cb94f]
Former-commit-id: 9a8854e0fa21ec59287649bd1d8f4716cbfd7f76