Make the story screen resize aware
These changes make the storyscreen resize aware by adding an event
handler to the UI component and giving it its own event context. On an
window-event it'll restart relayout the entire scene and re-start the
drawing of the text.

Known issue: Gamescreen buttons appear when you move over them. This is beyond the scope of this commit and will be fixed in other changes.


Former-commit-id: 0ee09651b33527a58b7e6a5ce0a8dac3d9ccf3bd
Former-commit-id: 86054f81a271f891bf863cd055e8a718aeba5c5a
Former-commit-id: c8167d5718fa4047ab83a3d780477dabffba088a
Former-commit-id: 5d6fc96da46f26cbccc2730f423535857ee690d7 [formerly d7243907ca0a4d03d0b8b4a4a6da812eab69df1c]
Former-commit-id: a48554f8fc80261303e73d34e42bfaa599af9835