Attempt to fix listbox scrollbar position after content changes.
Proviously, if a row was added/removed from a listbox, the scrolling
position of the listbox would be reset to 0. The expected behaviour is
that the visible part of the list does not change, this means that the
position is increased by the added space if the space was inserted
before the visible area, othewise it shouldn't be increased.

Unforunateley the code that adds/removes new rows does not know where
the extra space was inserted/removed because this is not calculated
until the next draw which calls layout_children(). So this doesn't work
very well, even after this commit.

Will try to fix this for treeviews later which might be easier because
treeviews don't use the tgenterator class and specially don't support
sorting of its elements.

We might want to revert the listbox.cpp/hpp part of this commit later
in case it doesn't work.


Former-commit-id: d87ad1b93c3209a320f8737038115ac1c237f05c
Former-commit-id: b8203a18b59876aa413b372a0fcdbc432867e256
Former-commit-id: beb40f03eb4ae93aef7749bad3a18241a15f1d87
Former-commit-id: e05866654d8c990731893866aa18e58292874c3f [formerly 4c5ecefb9f35b061d04a0009d8c398fb234f931f]
Former-commit-id: 8963e4eb980107144ef54c184402bd60909aed77