fix sides with no type= attribute
If you dont want that the engine uses the [side] wml table to create a
leader for that side, it is now enough to just omit the type attribute.
Previously it was necessary to add a no_leader=yes attribute.

The no_leader attribute is still used by the mp connect engine (which is
also used in sp) to prevent the mp connect engine from adding a random
leader from the current era. no_leader now defaults to yes for sides in
[scenario] and to no for sides in [multiplayer] so that the connect
engine now by default doesn't add a random mp leader to sides in sp
games.


Former-commit-id: 4d341026e71da88038e4647163a8837c80d7dd3c
Former-commit-id: f26504ffbdc945128a91f4de67a60ec9c7442d57
Former-commit-id: 5724694a373cee24052ba1a91a38174fe3aeda1c
Former-commit-id: 65847c90f45ae845abc6e3301cc772401da93682 [formerly 912c92d5da3f67ac30b278904d42760a04ca1f95]
Former-commit-id: af1a813d0e7abd5e95305dd3f8e943e4c7149aca