Add a dynamic layout functionality to the game display
This adds dynamic layout functionality, which is invoked on a full
redraw and upon element creation. It also makes sure that there's no
attempt to do a surface restore on negative widhts or heights as
introduced support for in SDL2. Finally, there's a workaround for text
being clipped by 1-2 pixels in GUI1 buttons.


Former-commit-id: 730d0fcac62319aca98bdec3ee70506ea11cf7a3
Former-commit-id: 40b167e6fcf135f8e15acb33bb29296cd90673ec
Former-commit-id: 93f055620be49c97ac4611a4c48351c0ed06a45b
Former-commit-id: a94eddb90c161e855259d77c70c479ff52011d0d [formerly c6925775e45de6b2a1067b2b65475b391e2d4663]
Former-commit-id: 7b09cccd5737cda6fcf89a892970d41a6da0dfb4