Make window events a separate handler for sdl_events
This adds a new function to sdl_handler, handle_window_event that is called for all window events. It is propogated to all event contexts instead of just the current one.
events::raise_draw_all_event (and a volatile variant) are added, that also sends draw events to all event contexts in reverse order so that the draw-stack can be redrawn bottom-up.

Former-commit-id: ae240bd31679ec2b56ef1a50a8c0995711673193
Former-commit-id: b87bb3e7090a454a6f281f64fcb6127eef24049c
Former-commit-id: 44a67cfb094b85c1dc72eb6499ef5efa0be5bb34
Former-commit-id: 8815121018df81e3b45bff68e998d81ae3e7c60e [formerly 1c9ad32c602a8320b76d52f798bbe9b41ec87b18]
Former-commit-id: 6fc0cd1702eefd1966cbf3663527607638b46713