remove game_display instance form game_launcher
This fixes https://gna.org/bugs/?24139 so that there is always at most
in display object. display objects are the ingame/editor views which
use theme wml and show the map with its units. So the exist only during
the editor or ingame.

This removes some display_manager objects which had no effect because the
registered display was a dummy.


Former-commit-id: 246cacb0ac92e52e3b94248decb0554a19ccfadb
Former-commit-id: a3eead60422c3917d58ade6bed0cd06b7d610723
Former-commit-id: e2664af58ab0faf35bf2d01669b0d4d23bc439ee
Former-commit-id: c82d9a11f1f580b7404fed59e7240a537bb62449 [formerly 9a9983d57b151245927735b45bf251efba0cd5f9]
Former-commit-id: dd00980d6312a1ea94c7c27463183e4a8c833d14