Add initial support for GUI layering
This introduces a subclass of events::sdl_handler that keeps track of
all instantiated classes. When a new instance is created it's pushed
onto a list of all tracked instances. When an instance is destroyed,
all known instances are sent a resize event (under SDL2) to mark them
as dirty. After that there's a DRAW_ALL event pushed onto the event
queue to cause a redraw. There's currently no logic in place to ensure
that there's only one DRAW_ALL event in the event queue at any given
time, but it should be harmless as long as no components are marked as
dirty.


Former-commit-id: ecc0745c7e9a91ba8e2c815ce61ceebffe57e812
Former-commit-id: b362e317c60a4ec0195f1518b3288d2f5330778b
Former-commit-id: 4006ce8e9f46df4e531f3cc3ade6b9f454289f62
Former-commit-id: c765a195d390d9ec3076dd39b02c90d9a62f73bd [formerly b267a9bb25c42cbbdfe784a49a9d2d7db46f1991]
Former-commit-id: 2d70bb68d1d17d0f7b4fd539a91841a39dab206c