added name= attribute to [item]
name= has no effect but can be used to delete the item. I also changed
[item] to automatically create an id (in case no id was given) and
return that (only usable if called from lua).

I used the id= field from overlay class for this, it was previously only
used by the editor. Not sure if it is a good idea to reuse the id field.


Former-commit-id: aff944d74c3d8d6f1490ee79dfa9dd479b921639
Former-commit-id: 76303d8c5cbfdec6879f2f01651f5d28136c0de3
Former-commit-id: 725db431b42accc653ab21e8e580295c1f2f0e08
Former-commit-id: 162a8ab15a43b511011cd105a20e5c65896749ea [formerly f36d5be79560774b1e13cc912276ede71aa9ae7b]
Former-commit-id: 33dc178f8124ede0b136c25dfb6877c7af18312b