Continued improvements to layered rendering
This makes the display class be aware of the layered rendering events (DRAW_ALL) and make it mark itself as dirty on certain window events. It also makes sure that the layers are redrawn in the correct order. GUI2 is now appearing to handle events properly, but GUI1 is still having issues with dialogs. And there's an assertion failure on program exit.


Former-commit-id: 6ae6ee07e132734b0ac308c311cbd52a95dcc73b
Former-commit-id: 28a524ac5e3a2c5b1857a30db752523c406c906b
Former-commit-id: 0145979a0384e5f91c6afe62b304cb78ec072701
Former-commit-id: d62a8f665d84ee8051a4293a30e6fe93f905654d [formerly a9f8fbc606dae0020389c7fcfb33374fff0a245b]
Former-commit-id: 2803cf3f47d029ad212751d9fcf1d1e40b0fb7c7