Fix event-context related assertion failure at exit
This fixes an assertion failure at exit which was caused by the event
contexts being cleaned up before the dummy display. There is no need
for the dummy display implementation to ever receive events so the
functionality has been extended for it to not join an event context,
resolving the assertion failure.


Former-commit-id: d468179bfd3e00e32f68c70eb0c049ba5b68aee9
Former-commit-id: 8fea794d075330049770e347200d7f4b3f103ec1
Former-commit-id: b0defaa9050be40a249ac384ee8ad1f1b8f56951
Former-commit-id: 022a7217be8fae516a9e56607833c5fd82094db8 [formerly 8c1fb691815cef75f493cd4701666ab315ff2024]
Former-commit-id: e042bde93c720c3fdc66186a07ca5bb163e9e3db