Add disabled flag to allow clearing of default hot-keys (Bugs #21983/#22218/#23981)
The original design of using a "null" command to indicate a disabled hot-key is ambiguous with regards to when to save a user hot-key to preferences as well as when a default hot-key should be flagged as disabled. So introduce a separate disabled flag to resolve the ambiguity.
Where the disabled flag is set to true, the hot-key should not be written to preferences unless it is a default hot-key.


Former-commit-id: 217cbc03e52f020bdd7cb8a43e1eed1fd44cdedc
Former-commit-id: 8481223befaba0612fb16112089d53b0cf6f596a
Former-commit-id: c408ff20cf1fbb262462a1d98c959fee1788bf27
Former-commit-id: 2ecafabe2eb3fc669b6e298d21c12d3d3fd035e9 [formerly eaa8cd0579ccc9c2973d66965bf506c565ff2a13]
Former-commit-id: b3b9c9db69b8c0de0ebf639c56930c8ea477bed0