added border to dark forecast map
this adds borders to the 2p dark forecast survival map, it also removes
the big empty space in the upper third of the map.

This also renames the mask files to map files, becasue they are actually
just normal maps, and the [terrain_mask] just replaces the current map
with these.


Former-commit-id: 376a8ec2fdc838c0ab2fdc984daf45fd0b90fa4e
Former-commit-id: 9247e78c42e736f098fae06aa72818e491859d47
Former-commit-id: 1aac641bc9149e0ee6463033f0b960e0154e84e8
Former-commit-id: 76beb29b1caae1b425d38a772a3fbcdc71fbeddc [formerly 8e60206ed21933055ac24ba3641d4fb2d2a3206a]
Former-commit-id: b320e0118232ae2ee808fea9098aa23f0a507e06