2p dark forecast: don't store unit pool types in savefiles
instead of storing the poolypesin savefiles we now have them as 'const'
lua array and store only the indexes to this array in the savefiles.
this has some advantages:
1) It makes the savefiles smaller
2) It simplyfies the lua code becasue we don't have to convert the wml
tables to lua tables anymore.

This also changes the scenario so that the spawned unit types are now
calculated at the beginning (they will be the same when reloading from a
savefile.)


Former-commit-id: 2a3b0f84ac6b07537dfe38b40862895d90e8e690
Former-commit-id: 10f436b8b9de83b8ead1c2752b3b2d832c7f5060
Former-commit-id: c4a2b76690afb6947ddda1813642bd65bcce0169
Former-commit-id: 41993ed68678a82e4349b696ea2d2a8f3cb05a32 [formerly 3691194a6eef31c561532a785beaeb1617ac85aa]
Former-commit-id: 8849446d3e09959d155b158be583c201871367a2