Make the MAKE_ENUM Doxygen documentation actually sane
Otherwise, most of the contents get attached to the bad_enum_cast type
definition instead.

Note that I couldn't figure out why the documentation for the stream
input operator says that twml_exception is thrown when in debug-mode
when the code does not actually do that at this time (the relevant
function call is commented out).


Former-commit-id: 6161d0a8ce5ddb7ebdc3f52a9766b908648f542d
Former-commit-id: 1989ae9e93e2f2d6542691d2fb1ee2c1eb97e183
Former-commit-id: fa2b8e7e42e711a2a0bce6d94863c84e22e7b168
Former-commit-id: 16e6808d4b78bec0ee4d125ca8e25c77ce6319d4 [formerly 02700c4fa85d2590c749ed7aa0cab4281c1119aa]
Former-commit-id: 49109ba692bd6a7759ae53076ce058fe8c38cbfc